mixcord v0.1.0 Mixcord.Api

Interface for Discord’s rest API.

Types

error()

error() :: {:error, Mixcord.Error.ApiError.t}

Represents a failed response from the API.

This occurs when hackney or HTTPoison fail, or when the API doesn’t respond with 200 or 204.

limit()

limit() :: Integer.t | :infinity

Represents a limit used to retrieve messages.

Integer number of messages, or :infinity to retrieve all messages.

locator()

locator ::
  {:before, Integer.t} |
  {:after, Integer.t} |
  {:around, Integer.t} |
  {}

Represents a tuple used to locate messages.

The first element of the tuple is an atom. The second element will be a message_id as an integer. The tuple can also be empty to search from the most recent message in the channel

status()

status() :: :dnd | :idle | :online | :invisible

Represents different statuses the bot can have.

:dnd - Red circle. :idle - Yellow circle. :online - Green circle. invisible - The bot will appear offline.

Functions

add_member(guild_id, user_id, options)

add_pinned_message(channel_id, message_id)

Pins a message.

Pins the message specified by message_id in the channel specified by channel_id.

Returns {:ok} if successful. error otherwise.

add_pinned_message!(channel_id, message_id)

Pins a message.

Pins the message specified by message_id in the channel specified by channel_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

bangify(to_bang)

batch_modify_guild_roles(guild_id, options)

begin_guild_prune(guild_id, options)

bulk_delete_messages(channel_id)

bulk_delete_messages!(channel_id)

create_channel(guild_id, options)

create_channel_invite(channel_id, options \\ %{})

create_channel_invite!(channel_id, options \\ %{})

create_guild_ban(guild_id, user_id, options)

create_guild_integrations(guild_id, options)

create_guild_roles(guild_id)

create_message(channel_id, content, tts \\ false)

create_message(Integer.t, String.t, boolean) ::
  error |
  {:ok, Mixcord.Map.Message.t}

create_message(Integer.t, Map.t, boolean) ::
  error |
  {:ok, Mixcord.Map.Message.t}

Send a message to a channel.

Send content to the channel identified with channel_id. Content can be either a binary containing the message you want to send or a Mixcord.Map.Embed map.

tts is an optional parameter that dictates whether the message should be played over text to speech.

Returns {:ok, Mixcord.Map.Message} if successful. error otherwise.

create_message!(channel_id, content, tts \\ false)

create_message!(Integer.t, String.t, boolean) ::
  no_return |
  Mixcord.Map.Message.t

Send a message to a channel.

Send content to the channel identified with channel_id. tts is an optional parameter that dictates whether the message should be played over text to speech.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns Mixcord.Map.Message if successful.

delete_channel(channel_id)

delete_channel!(channel_id)

delete_channel_permissions(channel_id, overwrite_id)

delete_channel_permissions!(channel_id, overwrite_id)

delete_guild(guild_id)

delete_guild_integrations(guild_id, integration_id)

delete_guild_role(guild_id, role_id)

delete_message(channel_id, message_id)

delete_message(Integer.t, Integer.t) :: error | {:ok}

Delete a message.

Delete a message specified by channel_id and message_id.

Returns {:ok} if successful. error otherwise.

delete_message!(channel_id, message_id)

delete_message!(String.t, Integer.t) :: no_return | {:ok}

Delete a message.

Delete a message specified by channel_id and message_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

delete_pinned_message(channel_id, message_id)

Unpins a message.

Unpins the message specified by message_id in the channel specified by channel_id.

Returns {:ok} if successful. error otherwise.

delete_pinned_message!(channel_id, message_id)

Unpins a message.

Unpins the message specified by message_id in the channel specified by channel_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

edit_channel(channel_id, options)

edit_channel!(channel_id, options)

edit_channel_permissions(channel_id, overwrite_id)

edit_channel_permissions!(channel_id, overwrite_id)

edit_guild(guild_id, options)

edit_message(channel_id, message_id, content)

edit_message(Integer.t, Integer.t, String.t) ::
  error |
  {:ok, Mixcord.Map.Message.t}

Edit a message.

Edit a message with the given content. Message to edit is specified by channel_id and message_id.

Returns the edited {:ok, Mixcord.Map.Message} if successful. error otherwise.

edit_message!(channel_id, message_id, content)

edit_message!(Integer.t, Integer.t, String.t) ::
  no_return |
  {:ok, Mixcord.Map.Message.t}

Edit a message.

Edit a message with the given content. Message to edit is specified by channel_id and message_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns the edited Mixcord.Map.Message if successful.

get_channel(channel_id)

get_channel!(channel_id)

get_channel_invites(channel_id)

get_channel_invites!(channel_id)

get_channel_message(channel_id, message_id)

get_channel_message!(channel_id, message_id)

get_channel_messages(channel_id, limit, locator)

get_channel_messages(Integer.t, limit, locator) ::
  error |
  {:ok, [Mixcord.Map.Message.t]}

Retrieve messages from a channel.

Retrieves limit number of messages from the channel with id channel_id. locator is a tuple indicating what messages you want to retrieve.

Returns {:ok, [Mixcord.Map.Message]} if successful. error otherwise.

get_channel_messages!(channel_id, limit, locator)

get_channel_messages_call(channel_id, limit, locator)

get_guild(guild_id)

get_guild_bans(guild_id)

get_guild_embed(guild_id)

get_guild_integrations(guild_id)

get_guild_invites(guild_id)

get_guild_prune(guild_id, options)

get_guild_roles(guild_id)

get_member(guild_id, user_id)

get_pinned_messages(channel_id)

get_pinned_messages!(channel_id)

get_token()

get_token() :: String.t

Returns the token of the bot.

get_voice_region(guild_id)

guild_members(guild_id, options)

modify_channel_position(guild_id, options)

modify_guild_embed(guild_id)

modify_guild_integrations(guild_id, integration_id, options)

modify_guild_roles(guild_id, role_id, options)

modify_member(guild_id, user_id, options)

remove_guild_ban(guild_id, user_id)

remove_member(guild_id, user_id)

request(method, route, body \\ "", options \\ [])

start_typing(channel_id)

Triggers the typing indicator.

Triggers the typing indicator in the channel specified by channel_id. The typing indicator lasts for about 8 seconds and then automatically stops.

Returns {:ok} if successful. error otherwise.

start_typing!(channel_id)

Triggers the typing indicator.

Triggers the typing indicator in the channel specified by channel_id. The typing indicator lasts for about 8 seconds and then automatically stops.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

sync_guild_integrations(guild_id, integration_id)

update_status(status, game)

update_status(status, String.t) :: no_return

Updates the status of the bot for all shards.

For more information see Mixcord.Api.update_status/3

update_status(pid, status, game)

update_status(Pid.t, status, String.t) :: no_return

Updates the status of the bot for a certain shard.

pid is the pid of the shard whose status you want to update. To update the status for all shards see Mixcord.Api.update_status/2 status is an atom that describes the status of the bot. See Mixcord.Api.status.t for available options. game is the text that will display ‘playing’ status of the game. This is the text below the bot’s name in the sidebar. Empty string will clear.