mixcord v0.1.0 Mixcord.Cache.Guild

Cache for guilds.

The ETS table name associated with the Guild Cache is :guilds. Besides the methods provided below you can call any other ETS methods on the table.

Example

info = :ets.info(:guilds)
[..., heir: :none, name: :guilds, size: 1, ...]
size = info[:size]
1

Functions

get(list)

get([{:id, integer}]) :: Mixcord.Map.Guild.t

get([{:message, Mixcord.Map.Message.t}]) :: Mixcord.Map.Guild.t

get!(list)