Elixir v0.1.0 Mixcord.Cache.UserCache

Cache for users.

The ETS table name associated with the User Cache is :users. Besides the methods provided below you can call any other ETS methods on the table.

Example

info = :ets.info(:users)
[..., heir: :none, name: :users, size: 1, ...]
size = info[:size]
1

Functions

get(list)

get([{:id, integer}]) ::
  {:ok, Mixcord.Struct.User.t} |
  {:error, atom}

Retrieves a user from the cache.

get/1 requires a keyword list as its only argument.

Returns {:ok, Mixcord.Struct.User.t} if found, {:error, atom} otherwise.

Example

case Mixcord.Cache.User.get(id: 1111222233334444) do
  {:ok, user} ->
    "We found " <> user.username
  {:error, _reason} ->
    "No es bueno"
end

get!(list)

get!([{:id, integer}]) :: Mixcord.Struct.User.t | no_return

Retrieves a user from the cache.

See get/1 for use and examples.

Returns Mixcord.Struct.User.t if found. Raises Mixcord.Error.CahceError if not found.

remove_struct_key(map)