mixcord v0.1.0 Mixcord.Constructs.Message

Struct representing a Discord message and various helper functions.

Functions

__struct__()

Defines the Message struct.

:id - String. Id of the message.

:channel_id - String. Id of the channel the message was sent in.

:author - Struct. A Mixcord.Constructs.User struct.

:content - String. Contents of the message.

:timestamp - Date. When this message was sent.

:edited_timestamp - ?Date. When this message was edited (or null if never).

:tts - Boolean. Whether this was a TTS message.

:mention_everyone - Boolean. Whether this message mentions everyone.

:mentions - List. A list of Mixcord.Constructs.User structs mentioned in the message.

:mention_roles - List. A list of Mixcord.Constructs.Role structs mentioned in this message.

:attachments - List. A list of attached files as maps.

:embeds - List. A list of embedded content as maps.

:nonce - ?String. Used for validating a message was sent.

:pinned - Boolean. Whether this message is pinned.

get_member()