mixcord v0.1.0 Mixcord.Constructs.TextChannel

Struct representing a Discord text channel and various helper functions.

Functions

__struct__()

Defines the Text Channel struct.

:id - String. The channel’s id.

:guild_id - String. Id of the channel’s guild.

:name - String. Name of the channel

:type - String. Text or Voice

:position - Integer. Ordering position of the channel

:is_private - Boolean. If channel is a DM

:permission_overwrites - List. List of overwrite objects as maps

:topic - String. Current channel topic.

:last_message_id - String. Id of the last message sent in the channel