mixcord v0.1.0 Mixcord.Map.Guild

Struct representing a Discord guild.

Types

afk_channel_id()

afk_channel_id() :: integer

The id of the guild’s afk channel

afk_timeout()

afk_timeout() :: integer

The time someone must be afk before being moved

channels()

channels() :: [Mixcord.Map.TextChannel.t]

List of channels

default_message_notifications()

default_message_notifications() :: integer

Level of verification

embed_channel_id()

embed_channel_id() :: integer

The id of the embedded channel

embed_enabled()

embed_enabled() :: boolean

Whether the guild is emeddable

emojis()

emojis() :: [Map.t]

List of emojis as maps

features()

features() :: [Map.t]

List of guild features

icon()

icon() :: String.t

The hash of the guild’s icon

id()

id() :: integer

The guild’s id

joined_at()

joined_at() :: String.t

Date the user joined the guild at

large()

large() :: boolean

Whether the guild is considered ‘large’

member_count()

member_count() :: integer

Total number of members in the guild

members()

members() :: [Mixcord.Map.Member.t]

List of members

mfa_level()

mfa_level() :: integer

Required MFA level of the guild

name()

name() :: String.t

The name of the guild.

owner_id()

owner_id() :: integer

The id of the guild owner

presences()

presences() :: [Map.t]

List of simple presence maps

region()

region() :: String.t

The id of the voice region

roles()

roles() :: [Mixcord.Map.Role.t]

List of roles

splash()

splash() :: String.t

The hash of the guild’s splash

t()

t() :: %Mixcord.Map.Guild{afk_channel_id: term, afk_timeout: term, channels: term, default_message_notifications: term, embed_channel_id: term, embed_enabled: term, emojis: term, features: term, icon: term, id: id, joined_at: term, large: term, member_count: term, members: term, mfa_level: term, name: term, owner_id: term, presences: term, region: term, roles: term, splash: term, unavailable: term, verification_level: term, voice_states: term}

unavailable()

unavailable() :: boolean

Whether the guild is avaliable

verification_level()

verification_level() :: integer

The level of verification

voice_states()

voice_states() :: [Map.t]

List of voice states as maps

Functions

__struct__()

Represents a Discord Guild.

:name - String. The guild’s name.

:icon - String. Hash of guild’s icon.

:splash - String. Hash of guild’s splash.

:owner_id - Integer. Id of the guild owner.

:region - String. Id of the voice region.

:afk_channel_id - Integer. Id of the guild’s afk channel.

:afk_timeout - Integer. How long someone must be afk before being moved.

:embed_enabled - Boolean. Is this guild embeddable.

:embed_channel_id - Integer. Id of the embedded channel.

:verification_level - Integer. Level of verification.

:default_message_notifications - Integer. Default message notification level.

:roles - List. List of Mixcord.Map.Role maps.

:emojis - List. List of emojis as maps.

:features - List. List of guild features.

:mfa_level - Integer. Required MFA level of the guild.

:joined_at - Date. Date the user joined the guild at.

:large - Boolean. Whether the guild is considered “large”.

:unavailable - Boolean. Whether the guild is available.

:member_count - Integer. Total number of members in the guild.

:voice_states - List. List of voice states as maps.

:members - List. List of Mixcord.Map.Member maps.

:channels - List. List of Mixcord.Map.Channel maps.

:presences - List. List of simple presence maps.