mixcord v0.1.0 Mixcord.Map.Member

Struct representing a Discord guild member.

User vs. Member

A user contains only general information about that user such as a username and an avatar. A member has everything that a user has, but also additional information on a per guild basis. This includes things like a nickname and a list of roles.

Types

deaf()

deaf() :: boolean

Whether the user is deafened

guild_id()

guild_id() :: integer

Id of the guild the member is part of

joined_at()

joined_at() :: String.t

Date the user joined the guild

mute()

mute() :: boolean

Whether the user is muted

nick()

nick() :: String.t | nil

The nickname of the user

roles()

roles() :: [Mixcord.Map.Role.t]

A list of role ids

t()

t() :: Map.t

user()

user() :: Mixcord.Map.User.t

The user struct

Functions

__struct__()

Represents a Discord Member.

:guild_id - String. Id of the guild the member is a part of.

:user - Struct. A Mixcord.Map.User struct.

:nick - ?String. Users guild nickname (if one is set).

:roles - List. A list of Mixcord.Map.Role ids.

:joined_at - Data. Date the user joined the guild.

:deaf - Boolean. If the user is deafened.

:mute - Boolean. If the user is muted.