mixcord v0.1.0 Mixcord.Map.Role

Struct representing a Discord role.

Types

color()

color :: integer

The hexadecimal color code

hoist()

hoist :: boolean

Whether the role is pinned in the user listing

id()

id :: integer

The id of the role

managed()

managed :: boolean

Whether the role is managed by an integration

mentionable()

mentionable :: boolean

Whether the role is mentionable

name()

name :: String.t

The name of the role

permissions()

permissions :: integer

The permission bit set

position()

position :: integer

The position of the role

t()

t :: Map.t

Functions

__struct__()

Represents a Discord Role.

:id - Integer. Id of the role.

:name - String. Name of the role.

:color - Integer. Integer representation of hexadecimal color code.

:hoist - Boolean. If the role is pinned in the user listing.

:position - Integer. Position of the role.

:permissions - Integer. Permission bit set.

:managed - Boolean. Whether this role is managed by an integration.

:mentionable - Boolean. Whether this role is mentionable.