mixcord v0.1.0 Mixcord.Rest.Client

Interface for Discord’s rest API.

Functions

create_message(channel_id, content, tts \\ false)

Specs

create_message(String.t, String.t, boolean) ::
  {:error, Map.t} |
  {:ok, Mixcord.Constructs.Message.t}

Send a message to a channel.

Send content to the channel identified with channel_id. tts is an optional parameter that dictates whether the message should be played over text to speech.

Returns {:ok, Mixcord.Constructs.Message} if successful. {:error, %{status_code: status_code, message: message}} otherwise.

create_message!(channel_id, content, tts \\ false)

Specs

create_message!(String.t, String.t, boolean) ::
  no_return |
  Mixcord.Constructs.Message.t

Send a message to a channel.

Send content to the channel identified with channel_id. tts is an optional parameter that dictates whether the message should be played over text to speech.

Raises Mixcord.Errors.ApiError if error occurs while making the rest call. Returns Mixcord.Constructs.Message if successful.

delete_message(channel_id, message_id)

Specs

delete_message(String.t, String.t) ::
  {:error, Map.t} |
  {:ok}

Delete a message.

Delete a message specified by channel_id and message_id.

Returns {:ok} if successful. {:error, %{status_code: status_code, message: message}} otherwise.

delete_message!(channel_id, message_id)

Specs

delete_message!(String.t, String.t) ::
  {:error, Map.t} |
  {:ok}

Delete a message.

Delete a message specified by channel_id and message_id.

Raises Mixcord.Errors.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

edit_message(channel_id, message_id, content)

Specs

edit_message(String.t, String.t, String.t) ::
  {:error, Map.t} |
  {:ok, Mixcord.Constructs.Message.t}

Edit a message.

Edit a message with the given content. Message to edit is specified by channel_id and message_id.

Returns the edited {:ok, Mixcord.Constructs.Message} if successful. {:error, %{status_code: status_code, message: message}} otherwise.

edit_message!(channel_id, message_id, content)

Specs

edit_message!(String.t, String.t, String.t) ::
  {:error, Map.t} |
  {:ok, Mixcord.Constructs.Message.t}

Edit a message.

Edit a message with the given content. Message to edit is specified by channel_id and message_id.

Raises Mixcord.Errors.ApiError if error occurs while making the rest call. Returns the edited Mixcord.Constructs.Message if successful.

init(token)

Specs

init(String.t) :: nil

Starts the rest client and initializes the agent containing your API token.

token()

Specs

token :: String.t

Returns the token of the bot.