Elixir v0.1.0 Mixcord.Shard.Dispatch.Consumer behaviour

Consumer process for gateway event handling.

Consuming Gateway Events

To handle events, Mixcord uses a GenStage implementation. GenStage is “new” with Elixir version 1.4, expanding on the old functionality of GenEvent.

Mixcord defines the producer in the GenStage design. To consume the events you must create at least one consumer process. For an example of this behaviour please see here.

It is generally recommended that you spawn a consumer per core. To find this number you can use System.schedulers_online/0.

Types

build_ban_remove()

build_ban_remove() :: {:GUILD_BAN_REMOVE, {Mixcord.Struct.User.t}, ws_state}

channel_create()

channel_create() :: {:CHANNEL_CREATE, {Mixcord.Struct.Channel.t}, ws_state}

channel_delete()

channel_delete() :: {:CHANNEL_DELETE, {Mixcord.Struct.Channel.t}, ws_state}

channel_pins_ack()

channel_pins_ack() :: {:CHANNEL_PINS_ACK, {Map.t}, ws_state}

channel_pins_update()

channel_pins_update() :: {:CHANNEL_PINS_UPDATE, {Map.t}, ws_state}

channel_update()

channel_update() :: {:CHANNEL_UPDATE, {old_channel :: Mixcord.Struct.Channel.t, new_channel :: Mixcord.Struct.Channel.t}, ws_state}

event()

event ::
  channel_create |
  channel_delete |
  channel_update |
  channel_pins_ack |
  channel_pins_update |
  guild_ban_add |
  build_ban_remove |
  guild_create |
  guild_update |
  guild_delete |
  guild_emojis_update |
  guild_integrations_update |
  guild_member_add |
  guild_members_chunk |
  guild_member_remove |
  guild_member_update |
  guild_role_create |
  guild_role_delete |
  guild_role_update |
  message_create |
  message_delete |
  message_delete_bulk |
  message_update |
  message_reaction_add |
  message_reaction_remove |
  presence_update |
  ready |
  resumed |
  typing_start |
  user_settings_update |
  user_update |
  voice_state_update |
  voice_server_update

from()

from() :: {pid, tag :: term}

Tuple describing the client of a call request.

pid is the PID of the caller and tag is a unique term used to identify the call.

guild_ban_add()

guild_ban_add() :: {:GUILD_BAN_ADD, {Mixcord.Struct.User.t}, ws_state}

guild_create()

guild_create() :: {:GUILD_CREATE, {new_guild :: Mixcord.Struct.Guild.t}, ws_state}

guild_delete()

guild_delete() :: {:GUILD_DELETE, {old_guild :: Mixcord.Struct.Guild.t}, ws_state}

guild_emojis_update()

guild_emojis_update() :: {:GUILD_EMOJIS_UPDATE, {old_emojis :: [Mixcord.Struct.Message.Emoji.t], new_emojis :: [Mixcord.Struct.Message.Emoji.t]}, ws_state}

guild_integrations_update()

guild_integrations_update() :: {:GUILD_INTEGERATIONS_UPDATE, {Map.t}, ws_state}

guild_member_add()

guild_member_add() :: {:GUILD_MEMBER_ADD, {new_member :: Mixcord.Struct.Guild.Member.t}, ws_state}

guild_member_remove()

guild_member_remove() :: {:GUILD_MEMBER_REMOVE, {old_member :: Mixcord.Struct.Guild.Member.t}, ws_state}

guild_member_update()

guild_member_update() :: {:GUILD_MEMBER_UPDATE, {old_member :: Mixcord.Struct.Guild.Member.t, new_member :: Mixcord.Struct.Guild.Member.t}, ws_state}

guild_members_chunk()

guild_members_chunk() :: {:GUILD_MEMBERS_CHUNK, {Map.t}, ws_state}

guild_role_create()

guild_role_create() :: {:GUILD_ROLE_CREATE, {new_role :: Mixcord.Struct.Guild.Role.t}, ws_state}

guild_role_delete()

guild_role_delete() :: {:GUILD_ROLE_DELETE, {old_role :: Mixcord.Struct.Guild.Role.t}, ws_state}

guild_role_update()

guild_role_update() :: {:GUILD_ROLE_UPDATE, {old_role :: Mixcord.Struct.Guild.Role.t, new_role :: Mixcord.Struct.Guild.Role.t}, ws_state}

guild_update()

guild_update() :: {:GUILD_CREATE, {old_guild :: Mixcord.Struct.Guild.t, new_guild :: Mixcord.Struct.Guild.t}, ws_state}

message_create()

message_create() :: {:MESSAGE_CREATE, {message :: Mixcord.Struct.Message.t}, ws_state}

message_delete()

message_delete() :: {:MESSAGE_DELETE, {message :: Mixcord.Struct.Message.t}, ws_state}

message_delete_bulk()

message_delete_bulk() :: {:MESSAGE_DELETE_BULK, {updated_messages :: [Mixcord.Struct.Message.t]}, ws_state}

message_reaction_add()

message_reaction_add() :: {:MESSAGE_REACTION_ADD, Map.t}

message_reaction_remove()

message_reaction_remove() :: {:MESSAGE_REACTION_REMOVE, Map.t}

message_update()

message_update() :: {:MESSAGE_UPDATE, {updated_message :: Mixcord.Struct.Message.t}, ws_state}

presence_update()

presence_update() :: {:PRESENCE_UPDATE, {Map.t}, ws_state}

ready()

ready() :: {:READY, {Map.t}, ws_state}

resumed()

resumed() :: {:RESUMED, {Map.t}, ws_state}

state()

state() :: Map.t

The state of the consumer process.

typing_start()

typing_start() :: {:TYPING_START, {Map.t}, ws_state}

user_settings_update()

user_settings_update() :: no_return

user_update()

user_update() :: {:USER_UPDATE, {old_user :: Mixcord.Struct.User.t, new_user :: Mixcord.Struct.User.t}, ws_state}

voice_server_update()

voice_server_update() :: {:VOICE_SERVER_UPDATE, {Map.t}, ws_state}

voice_state_update()

voice_state_update() :: {:VOICE_STATE_UPDATE, {Map.t}, ws_state}

ws_state()

ws_state() :: Map.t

The state of the websocket connection for the shard the event occured on.

Functions

start_link(mod)

Callbacks

handle_event(event, state)

handle_event(event, state) :: {:ok, Map.t}

Callback used to handle events.

Event is the event name as an atom, and ws_state is the current state of the websocket that the event was received on. For more information on this please see Mixcord.Shard.Payload.state_map.t.

from is the process information of the producer from which the demand was received. state is the internal state of your consumer.