mixcord v0.1.0 Mixcord.Shard.Dispatch.Consumer.Events behaviour

Defines the events you can handle in a consumer.

Types

build_ban_remove()

build_ban_remove() :: no_return

channel_create()

channel_create() :: {:CHANNEL_CREATE, Mixcord.Struct.Channel.t}

channel_delete()

channel_delete() :: {:CHANNEL_DELETE, Mixcord.Struct.Channel.t}

channel_pins_ack()

channel_pins_ack() :: no_return

channel_pins_update()

channel_pins_update() :: no_return

channel_update()

channel_update() :: {:CHANNEL_UPDATE, old_channel :: Mixcord.Struct.Channel.t, new_channel :: Mixcord.Struct.Channel.t}

event()

event ::
  channel_create |
  channel_delete |
  channel_update |
  channel_pins_ack |
  channel_pins_update |
  guild_ban_add |
  build_ban_remove |
  guild_create |
  guild_update |
  guild_delete |
  guild_emojis_update |
  guild_integrations_update |
  guild_member_add |
  guild_members_chunk |
  guild_member_remove |
  guild_member_update |
  guild_role_create |
  guild_role_delete |
  guild_role_update |
  message_create |
  message_delete |
  message_delete_bulk |
  message_update |
  presence_update |
  ready |
  resumed |
  typing_start |
  user_settings_update |
  user_update |
  voice_state_update |
  voice_server_update

from()

from() :: {pid, tag :: term}

Tuple describing the client of a call request.

pid is the PID of the caller and tag is a unique term used to identify the call.

guild_ban_add()

guild_ban_add() :: no_return

guild_create()

guild_create() :: {:GUILD_CREATE, old_guild :: Mixcord.Struct.Guild.t}

guild_delete()

guild_delete() :: no_return

guild_emojis_update()

guild_emojis_update() :: no_return

guild_integrations_update()

guild_integrations_update() :: no_return

guild_member_add()

guild_member_add() :: no_return

guild_member_remove()

guild_member_remove() :: no_return

guild_member_update()

guild_member_update() :: no_return

guild_members_chunk()

guild_members_chunk() :: no_return

guild_role_create()

guild_role_create() :: no_return

guild_role_delete()

guild_role_delete() :: no_return

guild_role_update()

guild_role_update() :: no_return

guild_update()

guild_update() :: no_return

message_create()

message_create() :: no_return

message_delete()

message_delete() :: no_return

message_delete_bulk()

message_delete_bulk() :: no_return

message_update()

message_update() :: no_return

presence_update()

presence_update() :: no_return

ready()

ready() :: no_return

resumed()

resumed() :: no_return

state()

state() :: term

typing_start()

typing_start() :: no_return

user_settings_update()

user_settings_update() :: no_return

user_update()

user_update() :: no_return

voice_server_update()

voice_server_update() :: no_return

voice_state_update()

voice_state_update() :: no_return

ws_state()

ws_state() :: state

Callbacks

handle_event({}, from, state)

handle_event({event, ws_state}, from, state) :: {:noreply, [], state}