Elixir v0.1.0 Mixcord.Struct.Guild.Member

Struct representing a Discord guild member.

User vs. Member

A user contains only general information about that user such as a username and an avatar. A member has everything that a user has, but also additional information on a per guild basis. This includes things like a nickname and a list of roles.

Types

deaf()

deaf() :: boolean

Whether the user is deafened

guild_id()

guild_id() :: integer

Id of the guild the member is part of

joined_at()

joined_at() :: String.t

Date the user joined the guild

mute()

mute() :: boolean

Whether the user is muted

nick()

nick() :: String.t | nil

The nickname of the user

roles()

roles() :: [Integer.t]

A list of role ids

t()

t() :: %Mixcord.Struct.Guild.Member{deaf: deaf, guild_id: guild_id, joined_at: joined_at, mute: mute, nick: nick, roles: roles, user: user}

user()

user() :: Mixcord.Struct.User.t

The user struct