mixcord v0.1.0 Mixcord.Struct.Member

Struct representing a Discord guild member.

User vs. Member

A user contains only general information about that user such as a username and an avatar. A member has everything that a user has, but also additional information on a per guild basis. This includes things like a nickname and a list of roles.

Functions

__struct__()

Defines the Member struct.

:guild_id - String. Id of the guild the member is a part of.

:user - Struct. A Mixcord.Struct.User struct.

:nick - ?String. Users guild nickname (if one is set).

:roles - List. A list of Mixcord.Struct.Role ids.

:joined_at - Data. Date the user joined the guild.

:deaf - Boolean. If the user is deafened.

:mute - Boolean. If the user is muted.