Elixir v0.1.0 Mixcord.Struct.User

Struct representing a Discord user.

User vs. Member

A user contains only general information about that user such as a username and an avatar. A member has everything that a user has, but also additional information on a per guild basis. This includes things like a nickname and a list of roles.

Types

avatar()

avatar() :: String.t

User’s avatar hash

bot()

bot() :: boolean

Whether the user is a bot

discriminator()

discriminator() :: String.t

The user’s 4—digit discord-tag

email()

email() :: String.t

The user’s email

id()

id() :: integer

The user’s id

mfa_enabled()

mfa_enabled() :: boolean

Whether the user has two factor enabled

t()

t() :: %Mixcord.Struct.User{avatar: avatar, bot: bot, discriminator: discriminator, email: email, id: id, mfa_enabled: mfa_enabled, username: username, verified: verified}

username()

username() :: String.t

The user’s username

verified()

verified() :: boolean

Whether the email on the account has been verified