Nostrum.Api.Webhook (Nostrum v0.10.0)

Functions for interacting with the Discord API's webhook endpoints.

See: https://discord.com/developers/docs/resources/webhook

Types

matrix()

@type matrix() :: m1() | m2() | m3()

Functions

create(channel_id, args, reason \\ nil)

@spec create(
  Nostrum.Struct.Channel.id(),
  %{name: String.t(), avatar: String.t()},
  Nostrum.Struct.Guild.AuditLogEntry.reason()
) :: Nostrum.Api.error() | {:ok, Nostrum.Struct.Webhook.t()}

Creates a webhook.

Parameters

channel_id - Id of the channel to send the message to.

args - Map with the following required keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.

delete(webhook_id, reason \\ nil)

@spec delete(Nostrum.Struct.Webhook.id(), Nostrum.Struct.Guild.AuditLogEntry.reason()) ::
  Nostrum.Api.error() | {:ok}

Deletes a webhook.

Parameters

webhook_id - Id of webhook to delete.

reason - An optional reason for the guild audit log.

edit_message(webhook_id, webhook_token, message_id, args)
(since 1.x.x)

@spec edit_message(
  Nostrum.Struct.Webhook.id(),
  Nostrum.Struct.Webhook.token(),
  Nostrum.Struct.Message.id(),
  map()
) :: Nostrum.Api.error() | {:ok, Nostrum.Struct.Message.t()}

Edits a message previously created by the same webhook, args are the same as execute/3, however all fields are optional.

execute(webhook_id, webhook_token, args, wait \\ false)

@spec execute(
  Nostrum.Struct.Webhook.id() | Nostrum.Struct.User.id(),
  Nostrum.Struct.Webhook.token() | Nostrum.Struct.Interaction.token(),
  matrix(),
  boolean()
) :: Nostrum.Api.error() | {:ok} | {:ok, Nostrum.Struct.Message.t()}

Executes a webhook.

Parameters

webhook_id - Id of the webhook to execute.

webhook_token - Token of the webhook to execute.

args - Map with the following allowed keys:

content - Message content.

files - List of Files to send.

embeds - List of embeds to send.

username - Overrides the default name of the webhook.

avatar_url - Overrides the default avatar of the webhook.

tts - Whether the message should be read over text to speech.

flags - Bitwise flags.

thread_id - Send a message to the specified thread within the webhook's channel.

allowed_mentions - Mentions to allow in the webhook message

wait - Whether to return an error or not. Defaults to false.

Note: If wait is true, this method will return a Message.t() on success.

At least one of content, files or embeds should be supplied in the args parameter.

execute_git(webhook_id, webhook_token, wait \\ false)

@spec execute_git(
  Nostrum.Struct.Webhook.id(),
  Nostrum.Struct.Webhook.token(),
  boolean()
) ::
  Nostrum.Api.error() | {:ok}

Executes a git webhook.

Parameters

webhook_id - Id of the webhook to execute.

webhook_token - Token of the webhook to execute.

execute_slack(webhook_id, webhook_token, wait \\ false)

@spec execute_slack(
  Nostrum.Struct.Webhook.id(),
  Nostrum.Struct.Webhook.token(),
  boolean()
) ::
  Nostrum.Api.error() | {:ok}

Executes a slack webhook.

Parameters

webhook_id - Id of the webhook to execute.

webhook_token - Token of the webhook to execute.

get(webhook_id)

@spec get(Nostrum.Struct.Webhook.id()) ::
  Nostrum.Api.error() | {:ok, Nostrum.Struct.Webhook.t()}

Gets a webhook by id.

Parameters

webhook_id - Id of the webhook to get.

get_message(webhook, message_id)
(since 1.x.x)

@spec get_message(Nostrum.Struct.Webhook.t(), Nostrum.Struct.Message.id()) ::
  Nostrum.Api.error() | {:ok, Nostrum.Struct.Message.t()}

Retrieves the original message of a webhook.

get_with_token(webhook_id, webhook_token)

@spec get_with_token(Nostrum.Struct.Webhook.id(), Nostrum.Struct.Webhook.token()) ::
  Nostrum.Api.error() | {:ok, Nostrum.Struct.Webhook.t()}

Gets a webhook by id and token.

This method is exactly like get/1 but does not require authentication.

Parameters

webhook_id - Id of the webhook to get.

webhook_token - Token of the webhook to get.

modify(webhook_id, args, reason \\ nil)

@spec modify(
  Nostrum.Struct.Webhook.id(),
  %{name: String.t(), avatar: String.t()},
  Nostrum.Struct.Guild.AuditLogEntry.reason()
) :: Nostrum.Api.error() | {:ok, Nostrum.Struct.Webhook.t()}

Modifies a webhook.

Parameters

webhook_id - Id of the webhook to modify.

args - Map with the following optional keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.

modify_with_token(webhook_id, webhook_token, args, reason \\ nil)

@spec modify_with_token(
  Nostrum.Struct.Webhook.id(),
  Nostrum.Struct.Webhook.token(),
  %{name: String.t(), avatar: String.t()},
  Nostrum.Struct.Guild.AuditLogEntry.reason()
) :: Nostrum.Api.error() | {:ok, Nostrum.Struct.Webhook.t()}

Modifies a webhook with a token.

This method is exactly like modify_webhook/1 but does not require authentication.

Parameters

webhook_id - Id of the webhook to modify.

webhook_token - Token of the webhook to get.

args - Map with the following optional keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.