Nostrum.Bot (nostrum v0.11.0-dev)

Supervisor for a bot along with all its associated components.

Usage

In your supervisor tree, write:

bot_options = %{
  consumer: MyBot.Consumer,
  # This is an example and includes privileged intents, but
  # should provide a useful set of intents for bots. Feel
  # free to adjust this as you see fit.
  intents: [
    :direct_messages,
    :guild_bans,
    :guild_members,
    :guild_message_reactions,
    :guild_messages,
    :guilds,
    :message_content
  ],
  wrapped_token: fn -> System.fetch_env!("BOT_TOKEN") end,
}
children = [
  {Nostrum.Bot, bot_options}
]

See bot_options/0 for which options you can configure here. You may also configure supervisor_options/0, which allow you to customize the options we start the Supervisor with by defining your child in the form {Nostrum.Bot, {bot_options, supervisor_options}}, but the default supervisor_options/0 should suit most needs.

Migration guide

If you are still using the old way of deploying nostrum, please do the following:

Remove the :nostrum, :token configuration option. nostrum will then assume that you are starting it as part of your own supervision tree, and automatic startup via the application is disabled.

Remove your consumer from your application supervisor tree.

In your supervisor tree, add the bot_options described above, and add Nostrum.Bot to your supervisor tree as described above. For the :intents setting, cut the value you previously had in your config.exs and use it here. Do the same with :num_shards, renaming it to :shards on the bot_options map. You can then remove the old config values.

In your consumer, change use Nostrum.Consumer to @behaviour Nostrum.Consumer.

In your consumer, add a default event handling clause for unmatched events at the end of the module, such as:

  def handle_event(_), do: :ok

Limitations

This module was introduced in version 0.11.0 to provide a first step towards migrating away from nostrum starting as an application. However, at present, you are not able to start multiple bots with this at the same time due to global names used internally. However, as we migrate away from the application-based deployment, further internal changes should not require any update from you.

Types

bot_options()
(since 0.11.0)

@type bot_options() :: %{
  :consumer => module(),
  :intents => :all | :nonprivileged | [atom()],
  :wrapped_token => (-> String.t()),
  optional(:shards) =>
    :auto
    | :manual
    | (num_shards :: pos_integer())
    | {lowest :: pos_integer(), highest :: pos_integer(),
       total :: pos_integer()}
}

Options to start a bot with.

Required fields

:consumer: A module implementing the Nostrum.Consumer behaviour that should be called for each event.

:intents: Which gateway intents to request. See the Gateway Intents documentation for more details.

:wrapped_token: A function that takes no arguments and returns the bot token to use. This is wrapped to prevent exposure of the token in stacktraces.

Optional fields

:shards: Shards that should be started with this bot. Possible values:

:auto uses the suggested amount of shards as provided by Discord. This value is the default.

:manual do not automatically spawn shards. In this case, it is your responsibility to spawn shards manually, see the manual sharding documentation.

pos_integer(): A number of shards to run. nostrum will warn if this is not the recommended amount.

{lowest, highest, total} starts shards lowest to highest. total should contain the total amount of shards that your bot is expected to have. Useful for splitting a single bot across multiple nodes, see the multi-node documentation for further information.

intents()
(since 0.11.0)

@type intents() :: :all | :nonprivileged | [atom()]

Which gateway intents to request.

A full reference can be found either on Discord or on our Gateway Intents documentation.

supervisor_options()
(since 0.11.0)

@type supervisor_options() :: [Supervisor.init_option()]

Options to pass to the bot's supervisor.

By default, we set the strategy: :one_for_one. No other options are set by nostrum.

Functions

child_spec(nostrum_options)
(since 0.11.0)

@spec child_spec(options()) :: Supervisor.child_spec()