Elixir v0.2.1 Nostrum.Cache.Guild.GuildServer

Module for interacting with Guild Servers.

Types

search_criteria()

search_criteria ::
  [{:id, integer}] |
  [{:channel_id, integer}] |
  Nostrum.Struct.Message.t

Represents different ways to search for a guild.

Examples

Search by Nostrum.Struct.Guild.id/0

  GuildServer.get(id: 666666666666666)

Search by Nostrum.Struct.Guild.TextChannel.id/0

  GuildServer.get(channel_id: 314159265358979)

Search via Nostrum.Struct.Message

  msg = Nostrum.Api.create_message!(1618033988731, "Drake confirmed British")
  GuildServer.get(msg)

Functions

get(search_criteria)

get(search_criteria) ::
  {:error, reason :: atom} |
  {:ok, Nostrum.Struct.Guild.t}

Retrieves a guild from the cache.

Parameters

search_criteria - The criteria used to search. See search_criteria/0 for more info.

Example

case Nostrum.Cache.Guild.GuildServer.get(id: 1234567891234789) do
  {:ok, guild} -> "Guild has #{length(guild.members)} members"
  {:error, _reason} -> "Guild is MIA"
end

get!(search_criteria)

get!(search_criteria) :: no_return | {Nostrum.Struct.Guild.t}

Retrieves a guild from the cache.

See Nostrum.Cache.Guild.GuildServer.get/1 for usage.

Raises Nostrum.Error.CacheError if unable to find the guild.

get_value(search_criteria, key)

get_value(search_criteria, key :: atom) ::
  {:error, reason :: atom} |
  {:ok, term}

Retrieves a value from a guild in the cache.

This method should be preferred over the Nostrum.Cache.Guild.GuildServer.get/1 method in the event that you need only one kv pair from the guild.

Parameters

search_criteria - The criteria used to search. See search_criteria/0 for more info.

key - The key to search for. See Nostrum.Struct.Guild for available keys.

Example

case Nostrum.Cache.Guild.Guildserver.get_value([id: 69696969696969], :name) do
  {:ok, sick_guild_name} -> "#{sick_guild_name} is sick"
  {:error, reason} -> "Whatchu' searchin' for fam?"
end

get_value!(search_critera, key)

get_value!(search_criteria, key :: atom) :: no_return | term

Retrieves a value from a guild in the cache.

See Nostrum.Cache.Guild.GuildServer.get_value!/2 for usage.

Raises Nostrum.Error.CacheError if unable to find the guild or key.