Nostrum.Cache.GuildCache.ETS (Nostrum v0.5.0)

An ETS-based cache for guilds.

The supervisor defined by this module will set up the ETS table associated with it.

The default table name under which guilds are cached is nostrum_guilds. In addition to the cache behaviour implementations provided by this module, you can also call regular ETS table methods on it, such as :ets.info.

Note that users should not call the functions not related to this specific implementation of the cache directly. Instead, call the functions of Nostrum.Cache.GuildCache directly, which will dispatch to the configured cache.

Functions

all()

Specs

all() :: Enum.t()

Retrieve all guilds from the cache.

channel_create(guild_id, channel)

Specs

channel_create(Nostrum.Struct.Guild.id(), map()) :: Nostrum.Struct.Channel.t()

Create the given channel for the given guild in the cache.

channel_delete(guild_id, channel_id)

Specs

channel_delete(Nostrum.Struct.Guild.id(), Nostrum.Struct.Channel.id()) ::
  Nostrum.Struct.Channel.t() | :noop

Delete the channel from the given guild in the cache.

channel_update(guild_id, channel)

Specs

channel_update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Channel.t(), Nostrum.Struct.Channel.t()}

Update the channel on the given guild in the cache.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

create(guild)

Specs

create(Nostrum.Struct.Guild.t()) :: true

Create the given guild in the cache.

delete(guild_id)

Specs

delete(Nostrum.Struct.Guild.id()) :: Nostrum.Struct.Guild.t() | nil

Delete the given guild from the cache.

emoji_update(guild_id, emojis)

Specs

emoji_update(Nostrum.Struct.Guild.id(), [map()]) ::
  {[Nostrum.Struct.Emoji.t()], [Nostrum.Struct.Emoji.t()]}

Update the emoji list for the given guild in the cache.

get(id)

Specs

get(Nostrum.Struct.Guild.id()) ::
  {:ok, Nostrum.Struct.Guild.t()} | {:error, Nostrum.Cache.GuildCache.reason()}

Retrieve a guild from the cache by ID.

get_by(clauses)

Specs

get_by(Nostrum.Cache.GuildCache.clauses()) ::
  {:ok, Nostrum.Struct.Guild.t()} | {:error, Nostrum.Cache.GuildCache.reason()}

Get a guild from the cache using the given selectors.

init(init_arg)

Set up the cache's ETS table.

member_add(guild_id, payload)

Specs

member_add(Nostrum.Struct.Guild.id(), map()) :: Nostrum.Struct.Guild.Member.t()

Add the given member to the given guild in the cache.

member_chunk(guild_id, member_chunk)

Create a chunk of members for the given guild in the cache.

member_remove(guild_id, user)

Specs

member_remove(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.t()} | :noop

Remove the given member from the given guild in the cache.

member_update(guild_id, member)

Specs

member_update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.t() | nil,
   Nostrum.Struct.Guild.Member.t()}

Update the given member for the given guild in the cache.

role_create(guild_id, role)

Specs

role_create(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.t()}

Create the given role in the given guild in the cache.

role_delete(guild_id, role_id)

Specs

role_delete(Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.id()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.t()} | :noop

Delete the given role from the given guild in the cache.

role_update(guild_id, role)

Specs

role_update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.t(),
   Nostrum.Struct.Guild.Role.t()}

Update the given role in the given guild in the cache.

select(id, selector)

Specs

select(Nostrum.Struct.Guild.id(), Nostrum.Cache.GuildCache.selector()) ::
  {:ok, any()} | {:error, Nostrum.Cache.GuildCache.reason()}

Select values from the guild with the matching ID.

select_all(selector)

Specs

select_all(Nostrum.Cache.GuildCache.selector()) :: Enum.t()

Retrieve all guilds matching the given selector from the cache.

select_by(clauses, selector)

Specs

select_by(
  Nostrum.Cache.GuildCache.clauses(),
  Nostrum.Cache.GuildCache.selector()
) :: {:ok, any()} | {:error, Nostrum.Cache.GuildCache.reason()}

Select values using a selector for a guild that matches the given clauses.

start_link(init_arg)

Start the supervisor.

tabname()

Specs

tabname() :: atom()

Retrieve the ETS table name used for the cache.

update(payload)

Specs

update(map()) :: {Nostrum.Struct.Guild.t(), Nostrum.Struct.Guild.t()}

Update the given guild in the cache.

voice_state_update(guild_id, payload)

Specs

voice_state_update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), [map()]}

Update guild voice states with the given voice state in the cache.