Elixir v0.2.1 Nostrum.Cache.Mapping.ChannelGuild

Mapping of channel -> guild relationships.

Given just the id of a guild channel, there’s no way to know what guild it belongs to. This mapping is an ETS table that stores this information.

Functions

get_guild(channel_id)

get_guild(integer) :: integer

Gets a guild_id from a channel_id.