Nostrum.Cache.MemberCache.Mnesia (Nostrum v0.9.0-alpha3)

An Mnesia-based cache for guild members.

Please note that this module is only compiled if Mnesia is available on your system. See the Mnesia section of the State documentation for more information.

To retrieve the table name used by this cache, use table/0.

Functions

bulk_create(guild_id, members)
(since 0.8.0)

Bulk create a chunk of members for the given guild in the cache.

child_spec(init_arg)
(since 0.8.0)

Returns a specification to start this module under a supervisor.

See Supervisor.

clear()
(since 0.8.0)

@spec clear() :: :ok

Clear any objects in the cache.

create(guild_id, payload)
(since 0.8.0)

@spec create(Nostrum.Struct.Guild.id(), map()) :: Nostrum.Struct.Guild.Member.t()

Add the given member to the given guild in the cache.

delete(guild_id, user_id)
(since 0.8.0)

@spec delete(Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.user_id()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.t()} | :noop

Remove the given member from the given guild in the cache.

init(init_arg)
(since 0.8.0)

Set up the cache's Mnesia table.

query_handle()
(since 0.8.0)

@spec query_handle() :: :qlc.query_handle()

Get a QLC query handle for the member cache.

start_link(init_arg)
(since 0.8.0)

Start the supervisor.

table()
(since 0.8.0)

@spec table() :: atom()

Retrieve the Mnesia table name used for the cache.

teardown()
(since 0.8.0)

@spec teardown() :: {:atomic, :ok} | {:aborted, term()}

Drop the table used for caching.

update(guild_id, payload)
(since 0.8.0)

@spec update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.t() | nil,
   Nostrum.Struct.Guild.Member.t()}

Update the given member for the given guild in the cache.

wrap_qlc(fun)
(since 0.8.0)

Wrap QLC operations in a transaction.