Nostrum.Cache.PresenceCache.ETS (Nostrum v0.5.0)

ETS-based cache for user presences.

The ETS table name associated with the User Cache is presences. Besides the methods provided below you can call any other ETS methods on the table. If you need to access the name of the presence cache's table programmatically, use the tabname/0 function instead of hardcoding it in your application.

Example

info = :ets.info(presences)
[..., heir: :none, name: presences, size: 1, ...]
size = info[:size]
1

Functions

bulk_create(guild_id, presences)

Specs

bulk_create(Nostrum.Struct.Guild.id(), [map()]) :: :ok

Bulk create multiple presences in the cache.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

create(presence)

Specs

create(map()) :: :ok

Add the given presence data to the cache.

get(user_id, guild_id)

Specs

get(Nostrum.Struct.User.id(), Nostrum.Struct.Guild.id()) ::
  {:error, :presence_not_found} | {:ok, map()}

Retrieves a presence for a user from the cache by guild and id.

init(init_arg)

Set up the cache's ETS table.

start_link(init_arg)

Start the supervisor.

tabname()

Specs

tabname() :: atom()

Return the ETS table name used for this cache.

update(presence)

Specs

update(map()) :: {Nostrum.Struct.Guild.id(), nil | map(), map()} | :noop

Update the given presence data in the cache.