Nostrum.Cache.PresenceCache.Mnesia (Nostrum v0.9.0-alpha3)

An Mnesia-based cache for presences.

Please note that this module is only compiled if Mnesia is available on your system. See the Mnesia section of the State documentation for more information.

To retrieve the table name used by this cache, use table/0.

Functions

bulk_create(guild_id, presences)
(since 0.8.0)

Bulk create a chunk of presences for the given guild in the cache.

child_spec(init_arg)
(since 0.8.0)

Returns a specification to start this module under a supervisor.

See Supervisor.

clear()
(since 0.8.0)

@spec clear() :: {:atomic, :ok} | {:aborted, term()}

Clear any objects in the cache.

create(presence)
(since 0.8.0)

@spec create(map()) :: :ok

Add the given presence to the cache.

init(init_arg)
(since 0.8.0)

Set up the cache's Mnesia table.

query_handle()
(since 0.8.0)

@spec query_handle() :: :qlc.query_handle()

Retrieve a query handle for the table.

start_link(init_arg)
(since 0.8.0)

Start the supervisor.

table()
(since 0.8.0)

@spec table() :: atom()

Retrieve the table name used by the cache.

teardown()
(since 0.8.0)

@spec teardown() :: {:atomic, :ok} | {:aborted, term()}

Drop the table used for caching.

update(new)
(since 0.8.0)

@spec update(map()) :: {Nostrum.Struct.Guild.id(), presence | nil, presence} | :noop
when presence: Nostrum.Cache.PresenceCache.presence()

Update the given presence in the cache.

wrap_qlc(fun)
(since 0.8.0)

Wrap QLC operations in a transaction.