Nostrum.Cache.PresenceCache.NoOp (Nostrum v0.5.0-rc2)

A cache module that does not cache anything.

Useful for bots that do not need presence caching.

Functions

bulk_create(guild_id, presences)

Specs

bulk_create(Nostrum.Struct.Guild.id(), [map()]) :: :ok

Do not bulk create multiple presences in the cache.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

create(presence)

Specs

create(map()) :: :ok

Do not add the given presence data to the cache.

init(init_arg)

Start up the cache supervisor.

start_link(init_arg)

Start the supervisor.

update(presence)

Specs

update(map()) :: {Nostrum.Struct.Guild.id(), nil | map(), map()} | :noop

Return the presence update for consumers.