Nostrum.Cache.UserCache.Mnesia (nostrum v0.11.0-dev)

An Mnesia-based cache for users.

Please note that this module is only compiled if Mnesia is available on your system. See the Mnesia section of the State documentation for more information.

To retrieve the table name used by this cache, use table/0.

Functions

bulk_create(users)
(since 0.8.0)

Bulk create a chunk of users in the cache.

child_spec(init_arg)
(since 0.8.0)

Returns a specification to start this module under a supervisor.

See Supervisor.

clear()
(since 0.8.0)

@spec clear() :: :ok

Clear any objects in the cache.

create(payload)
(since 0.8.0)

@spec create(map()) :: Nostrum.Struct.User.t()

Create a user from upstream data.

delete(user_id)
(since 0.8.0)

@spec delete(Nostrum.Struct.User.id()) :: Nostrum.Struct.User.t() | :noop

Remove the given user from cache.

get(user_id)
(since 0.8.0)

@spec get(Nostrum.Struct.User.id()) ::
  {:ok, Nostrum.Struct.User.t()} | {:error, :user_not_found}

Retrieve a user from the cache.

init(init_arg)
(since 0.8.0)

Set up the cache's Mnesia table.

start_link(init_arg)
(since 0.8.0)

Start the supervisor.

table()
(since 0.8.0)

@spec table() :: atom()

Retrieve the Mnesia table name used for the cache.

teardown()
(since 0.8.0)

@spec teardown() :: {:atomic, :ok} | {:aborted, term()}

Drop the table used for caching.

update(payload)
(since 0.8.0)

@spec update(map()) :: {Nostrum.Struct.User.t() | nil, Nostrum.Struct.User.t()}

Update a User if it exists in the cache.

wrap_query(fun)
(since 0.8.0)

Wrap QLC operations in a transaction.