Nostrum.Shard.Session.Compression.Zlib (nostrum v0.11.0-dev)

Implementation of compression methods for the zlib compression algorithm.

Internal module

This module is intended for exclusive usage inside of nostrum, and is documented for completeness and people curious to look behind the covers.

Functions

create_context()

@spec create_context() :: :zlib.zstream()

Callback implementation for Nostrum.Shard.Session.Compression.create_context/0.

inflate(ctx, frame)

@spec inflate(:zlib.zstream(), iodata()) :: iolist()

Callback implementation for Nostrum.Shard.Session.Compression.inflate/2.

reset_context(ctx)

@spec reset_context(:zlib.zstream()) :: :zlib.zstream()

Callback implementation for Nostrum.Shard.Session.Compression.reset_context/1.