Nostrum.Store.GuildShardMapping.Mnesia (Nostrum v0.9.0)

Maintains a mapping of guild IDs to their shard numbers using Mnesia.

Please note that this module is only compiled if Mnesia is available on your system. See the Mnesia section of the State documentation for more information.

To retrieve the table name used by this cache, use table/0.

Functions

child_spec(init_arg)
(since 0.8.0)

Returns a specification to start this module under a supervisor.

See Supervisor.

create(guild_id, shard_num)
(since 0.8.0)

@spec create(Nostrum.Struct.Guild.id(), Nostrum.Struct.WSState.shard_num()) :: :ok

Create a new mapping for the given guild ID to the given shard ID.

delete(guild_id)
(since 0.8.0)

@spec delete(Nostrum.Struct.Guild.id()) :: :ok

Delete any stored mapping for the given guild ID.

get(guild_id)
(since 0.8.0)

@spec get(Nostrum.Struct.Guild.id()) :: Nostrum.Struct.WSState.shard_num() | nil

Get the shard number for the given guild ID.

init(init_arg)
(since 0.8.0)

Set up the store's Mnesia table.

start_link(init_arg)
(since 0.8.0)

Start the supervisor.

table()
(since 0.8.0)

@spec table() :: atom()

Retrieve the Mnesia table name used for the store.

teardown()
(since 0.8.0)

@spec teardown() :: {:atomic, :ok} | {:aborted, term()}

Drop the table used for the store.