Nostrum.Store.RatelimitBucket.ETS (Nostrum v0.7.0)

Stores ratelimit buckets via :ets.

If programmatic access to the ETS table is needed, please use the table/0 function.

Please do not use this module directly, apart from special functions such as table/0. Use Nostrum.Store.RatelimitBucket to call the configured mapping instead.

Functions

child_spec(init_arg)
(since 0.8.0)

Returns a specification to start this module under a supervisor.

See Supervisor.

cleanup(age)
(since 0.8.0)

@spec cleanup(pos_integer()) :: non_neg_integer()

Clean up bucket entries older than age milliseconds.

init(init_arg)
(since 0.8.0)

Set up the store's ETS table.

lookup(route)
(since 0.8.0)

@spec lookup(Nostrum.Store.RatelimitBucket.route()) ::
  Nostrum.Store.RatelimitBucket.bucket() | nil

Look up the most relevant ratelimiter bucket for the given route.

start_link(init_arg)
(since 0.8.0)

Start the supervisor.

table()
(since 0.8.0)

@spec table() :: :ets.table()

Retrieve the ETS table reference used for the store.

update(route, remaining)
(since 0.8.0)

@spec update(
  Nostrum.Store.RatelimitBucket.route(),
  Nostrum.Store.RatelimitBucket.remaining()
) ::
  boolean()

Update an existing routes remaining calls.

update(route, remaining, reset_time, latency)
(since 0.8.0)

@spec update(
  Nostrum.Store.RatelimitBucket.route(),
  Nostrum.Store.RatelimitBucket.remaining(),
  Nostrum.Store.RatelimitBucket.reset_time(),
  Nostrum.Store.RatelimitBucket.latency()
) :: true

Set up a new ratelimiter bucket from the given arguments.