Nostrum.Struct.ApplicationCommandInteractionData (Nostrum v0.5.0)

Struct for interaction data.

Types

component_type()
(since 0.5.0)

Specs

component_type() :: integer() | nil

For components, the type of the component

custom_id()
(since 0.5.0)

Specs

custom_id() :: String.t() | nil

For components, the custom_id of the component

id()

Specs

id() :: Nostrum.Snowflake.t() | nil

ID of the invoked command

interaction_type()
(since 0.5.0)

Specs

interaction_type() :: integer() | nil

The type of application command invoked. Official reference: https://discord.com/developers/docs/interactions/application-commands#application-command-object-application-command-types

name()

Specs

name() :: String.t() | nil

Name of the invoked command

options()

Specs

options() :: Nostrum.Struct.ApplicationCommandInteractionDataOption.t() | nil

Parameters and values supplied by the user, if applicable

resolved()
(since 0.5.0)

Specs

resolved() :: Nostrum.Struct.ApplicationCommandInteractionDataResolved.t() | nil

Converted users & roles & channels

select_values()
(since 0.5.0)

Specs

select_values() :: [String.t()] | nil

For select menu components, this will be a list of the values the user selected.

t()

Specs

t() :: %Nostrum.Struct.ApplicationCommandInteractionData{
  component_type: component_type(),
  custom_id: custom_id(),
  id: id(),
  name: name(),
  options: options(),
  resolved: resolved(),
  target_id: target_id(),
  type: interaction_type(),
  values: select_values()
}

Command interaction data for slash commands.

Used as part of Nostrum.Struct.Interaction.t/0.

Official reference: https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object

target_id()
(since 0.5.0)

Specs

target_id() :: Nostrum.Snowflake.t() | nil

ID of the user or message targeted by a context menu command