Nostrum.Struct.ApplicationCommandInteractionDataResolved (Nostrum v0.5.0)

Converted interaction payload.

Types

channels()

Specs

channels() ::
  %{
    required(Nostrum.Struct.Channel.id()) =>
      Nostrum.Struct.Channel.guild_text_channel()
  }
  | nil

IDs and corresponding partial channels.

The channels in this map only have the following keys set:

id

name

type

permissions

members()

Specs

members() ::
  %{required(Nostrum.Struct.User.id()) => Nostrum.Struct.Guild.Member.t()} | nil

IDs and corresponding partial members.

These members are missing values on the following fields:

user

deaf

mute

The corresponding user data can be looked up in users. For members that are part of this map, data for the corresponding user will always be included.

messages()

Specs

messages() ::
  %{required(Nostrum.Struct.Message.id()) => Nostrum.Struct.Message.t()} | nil

The IDs and corresponding messages.

roles()

Specs

roles() ::
  %{required(Nostrum.Struct.Guild.Role.id()) => Nostrum.Struct.Guild.Role.t()}
  | nil

IDs and corresponding roles

t()

Specs

t() :: %Nostrum.Struct.ApplicationCommandInteractionDataResolved{
  channels: channels(),
  members: members(),
  messages: messages(),
  roles: roles(),
  users: users()
}

Resolved interaction data

users()

Specs

users() ::
  %{required(Nostrum.Struct.User.id()) => Nostrum.Struct.User.t()} | nil

IDs and corresponding users