Nostrum.Struct.AutoModerationRule (Nostrum v0.9.0-alpha3)

Struct representing an auto-moderation rule.

Types

actions()
(since 0.7.0)

@type actions() :: [Nostrum.Struct.AutoModerationRule.Action.t()]

A list of Actions which will be performed when the rule is triggered.

creator_id()
(since 0.7.0)

@type creator_id() :: Nostrum.Struct.User.id()

The id of the user who created the rule

enabled()
(since 0.7.0)

@type enabled() :: boolean()

If the rule is enabled or not.

event_type()
(since 0.7.0)

@type event_type() :: 1

Indicates in what event context a rule should be checked

value	type	description

----	----	-----------

1	MESSAGE_SEND	when a member sends or edits a message in a guild

exempt_channels()
(since 0.7.0)

@type exempt_channels() :: [Nostrum.Struct.Channel.id()]

A list of channels that are exempt from the rule

exempt_roles()
(since 0.7.0)

@type exempt_roles() :: [Nostrum.Struct.Guild.Role.id()]

A list of roles that are exempt from the rule.

guild_id()
(since 0.7.0)

@type guild_id() :: Nostrum.Struct.Guild.id()

The id of the guild the rule belongs to

id()
(since 0.7.0)

@type id() :: Nostrum.Snowflake.t()

The id of the auto-moderation rule

name()
(since 0.7.0)

@type name() :: String.t()

The name of the rule

preset_values()
(since 0.7.0)

@type preset_values() :: 1..3

Values which represent the different presets defined by Discord

value	type	description

----	----	-----------

1	PROFANITY	Words which may be considered profane

2	HARMFUL_LINK	Words that refer to sexually explicit behavior or activity

3	SLURS	Personal insults or words that may be considered hate speech

t()
(since 0.7.0)

@type t() :: %Nostrum.Struct.AutoModerationRule{
  actions: actions(),
  creator_id: creator_id(),
  enabled: enabled(),
  event_type: event_type(),
  exempt_channels: exempt_channels(),
  exempt_roles: exempt_roles(),
  guild_id: guild_id(),
  id: id(),
  name: name(),
  trigger_metadata: trigger_metadata(),
  trigger_type: trigger_type()
}

trigger_metadata()
(since 0.7.0)

@type trigger_metadata() :: Nostrum.Struct.AutoModerationRule.TriggerMetadata.t()

Additional data used to determine if the rule should triggered.

The trigger_type/0 field will determine which of the following fields are present.

key	associated trigger_type

----	-----------

keywords	​KEYWORD

preset	KEYWORD_PRESET

trigger_type()
(since 0.7.0)

@type trigger_type() :: 1..4

Characters the type of content which triggered the rule

value	type	max per guild	description

----	----	-----	-----------

1	​KEYWORD	3	check if content contains words from a user defined list of keywords

2	HARMFUL_LINK	1	check if the content contains any harmful links

3	SPAM	1	check if the content represents generic spam

4	KEYWORD_PRESET	1	check if the content contains a list of discord defined keywords

note: HARMFUL_LINK and SPAM are not yet offically released at the time of this writing.