Nostrum.Struct.AutoModerationRule.Action (nostrum v0.11.0-dev)

Defines an action to be taken when a rule is triggered.

Types

action_type()
(since 0.7.0)

@type action_type() :: 1..4

The type of action to be taken.

value	action	description

----	----	-----------

1	BLOCK_MESSAGE	Blocks the message from being created

2	SEND_ALERT_MESSAGE	Logs the content of the message in the specified channel

3	TIMEOUT	timeout a user for a specified duration

4	BLOCK_MEMBER_INTERACTION	prevents a member from using text, voice, or other interactions

metadata()
(since 0.7.0)

@type metadata() :: Nostrum.Struct.AutoModerationRule.ActionMetadata.t()

t()
(since 0.7.0)

@type t() :: %Nostrum.Struct.AutoModerationRule.Action{
  metadata: metadata(),
  type: action_type()
}