Nostrum.Struct.AutoModerationRule.ActionMetadata (Nostrum v0.9.0-alpha3)

Struct representing any additional data used when an action is taken.

Types

send_alert_message_metadata()
(since 0.7.0)

@type send_alert_message_metadata() ::
  %Nostrum.Struct.AutoModerationRule.ActionMetadata{
    channel_id: Nostrum.Snowflake.t(),
    duration_seconds: term()
  }

The id of the channel to send an alert message to.

t()
(since 0.7.0)

@type t() :: send_alert_message_metadata() | timeout_metadata()

The type of metadata present depends on the action type.

value	type

----	----

channel_id	SEND_ALERT_MESSAGE

duration_seconds	TIMEOUT

timeout_metadata()
(since 0.7.0)

@type timeout_metadata() :: %Nostrum.Struct.AutoModerationRule.ActionMetadata{
  channel_id: term(),
  duration_seconds: 1..2_419_200
}

The number of seconds to timeout the user for, has a maximum of 2419200 seconds (4 weeks).