Nostrum.Struct.AutoModerationRule.TriggerMetadata (Nostrum v0.9.0-alpha3)

Struct representing the metadata of a trigger.

Types

keyword_metadata()
(since 0.7.0)

@type keyword_metadata() :: %Nostrum.Struct.AutoModerationRule.TriggerMetadata{
  keyword_filter: [String.t()],
  presets: term()
}

Contains the list of keywords to that will trigger the rule.

preset_value_metadata()
(since 0.7.0)

@type preset_value_metadata() :: %Nostrum.Struct.AutoModerationRule.TriggerMetadata{
  keyword_filter: term(),
  presets: [1..3]
}

A list of Values which represent the different presets defined by Discord

value	type	description

----	----	-----------

1	PROFANITY	Words which may be considered profane

2	HARMFUL_LINK	Words that refer to sexually explicit behavior or activity

3	SLURS	Personal insults or words that may be considered hate speech

t()
(since 0.7.0)

@type t() :: preset_value_metadata() | keyword_metadata()

Additional data used to determine if the rule should triggered.

The Nostrum.Struct.AutoModerationRule.trigger_type/0 of the parent struct determine which of the following fields are not nil.

key	associated trigger_type

----	-----------

keywords	​KEYWORD

preset	KEYWORD_PRESET