Nostrum.Struct.Component.ActionRow (Nostrum v0.5.0-rc2)

Action Rows.

Types

t()

Specs

t() :: %{
  type: Nostrum.Struct.Component.type(),
  components: [Nostrum.Struct.Component.components()]
}

Functions

action_row(opts \\ [])

Create an empty action row.

Options can be passed as a keyword list. The only supported option is a list of inner components

append(action_row, button)

Appends a button to the action row.

Returns the action row unchanged if there are already 5 buttons or if the action row contains a select menu.

append_lazy(action_row, button)

Lazily appends a button to the action row.

If there are already 5 buttons, the first one will be dropped.

flatten(map)

put(component, select_menu)

Puts the given component into the action row, any existing components are disgarded.

put_new(component, list_of_components)

Puts the components into the action row unless a list of inner components already exists.