Nostrum.Struct.Component.Button (Nostrum v0.5.0)

Helpers for dealing with Button Components

Types

interaction_button()

Specs

interaction_button() :: %Nostrum.Struct.Component{
  components: term(),
  custom_id: Nostrum.Struct.Component.custom_id(),
  disabled: Nostrum.Struct.Component.disabled(),
  emoji: Nostrum.Struct.Component.emoji(),
  label: Nostrum.Struct.Component.label(),
  max_length: term(),
  max_values: term(),
  min_length: term(),
  min_values: term(),
  options: term(),
  placeholder: term(),
  required: term(),
  style: Nostrum.Struct.Component.style(),
  type: Nostrum.Struct.Component.type(),
  url: term(),
  value: term()
}

link_button()

Specs

link_button() :: %Nostrum.Struct.Component{
  components: term(),
  custom_id: term(),
  disabled: Nostrum.Struct.Component.disabled(),
  emoji: Nostrum.Struct.Component.emoji(),
  label: Nostrum.Struct.Component.label(),
  max_length: term(),
  max_values: term(),
  min_length: term(),
  min_values: term(),
  options: term(),
  placeholder: term(),
  required: term(),
  style: Nostrum.Struct.Component.style(),
  type: Nostrum.Struct.Component.type(),
  url: Nostrum.Struct.Component.url(),
  value: term()
}

opt()

Specs

opt() ::
  {:style, Nostrum.Struct.Component.style()}
  | {:label, Nostrum.Struct.Component.label()}
  | {:emoji, Nostrum.Struct.Component.emoji()}
  | {:custom_id, Nostrum.Struct.Component.custom_id()}
  | {:disabled, Nostrum.Struct.Component.disabled()}
  | {:url, Nostrum.Struct.Component.url()}

opts()

Specs

opts() :: [opt()]

t()

Specs

t() :: link_button() | interaction_button()

Functions

button(opts \\ [])

Create a button.

This function provides direct access to the new/1 callback.

Note: While using this function directly you are not guaranteed to return a valid button, providing a valid combination of options becomes the responsibility of the reader. It is instead recommended to use the helper functions provided below.

disable(button)

Disables the button.

enable(button)

Enables the button.

flatten(map)

interaction_button(label, custom_id, opts \\ [])

Create an interaction button.

Read more about interaction buttons in the Nostrum.Struct.Component documentation.

link_button(label, url, opts \\ [])

Create a link button.

Read more about link buttons in the Nostrum.Struct.Component documentation.

put_style(button, style)

Changes the style of the button.

toggle(button)

Toggle the buttons disabled state.