Nostrum.Struct.Component.SelectMenu (Nostrum v0.5.0-rc2)

Select Menu

Types

opt()

Specs

opt() ::
  {:custom_id, Nostrum.Struct.Component.custom_id()}
  | {:options, Nostrum.Struct.Component.options()}
  | {:placeholder, Nostrum.Struct.Component.placeholder()}
  | {:min_values, Nostrum.Struct.Component.min_values()}
  | {:max_values, Nostrum.Struct.Component.max_values()}
  | {:disabled, Nostrum.Struct.Component.disabled()}

opts()

Specs

opts() :: [opt()]

t()

Specs

t() :: %{
  type: Nostrum.Struct.Component.type(),
  custom_id: Nostrum.Struct.Component.custom_id(),
  disabled: Nostrum.Struct.Component.disabled(),
  options: Nostrum.Struct.Component.options(),
  placeholder: Nostrum.Struct.Component.placeholder(),
  min_values: Nostrum.Struct.Component.min_values(),
  max_values: Nostrum.Struct.Component.max_values()
}

Functions

flatten(map)

select_menu(custom_id, opts \\ [])

Creates a select menu that can be used inside an action row.

Options can be passed as a keyword list.

Parameters

custom_id - lower case string, used for matching against when your application receives an interaction.

Options

disabled - If the select should be disabled

options - A list of options for the select menu, see Nostrum.Struct.Component.Option

placeholder - Value to be shown before anything is selected

min_values - minimum number of values the user must select, between 0 and 25, default is 1

max_values - maximum number of values the user must select, between 0 and 25, default is 1