Nostrum.Struct.Event.GuildBanRemove (Nostrum v0.5.1)

Sent when a user is unbanned from a guild

Types

guild_id()

Specs

guild_id() :: Nostrum.Struct.Guild.id()

ID of the guild

t()

Specs

t() :: %Nostrum.Struct.Event.GuildBanRemove{guild_id: guild_id(), user: user()}

Event sent when a user is unbanned from a guild

user()

Specs

user() :: Nostrum.Struct.User.t()

Unbanned user