Nostrum v0.4.5 Nostrum.Struct.Event.InviteCreate

Struct representing an Invite Create event

Types

channel_id()

channel_id() :: Nostrum.Struct.Channel.id()

Channel id of the channel this invite is for.

code()

code() :: String.t()

The unique invite code.

created_at()

created_at() :: String.t()

The time at which the invite was created.

guild_id()

guild_id() :: Nostrum.Struct.Guild.id() | nil

Guild id of the guild this invite is for.

inviter()

inviter() :: Nostrum.Struct.User.t() | nil

The user that created the invite.

max_age()

max_age() :: integer()

Duration (in seconds) after which the invite expires.

max_uses()

max_uses() :: integer()

Max number of times this invite can be used.

t()

t() :: %Nostrum.Struct.Event.InviteCreate{
  channel_id: channel_id(),
  code: code(),
  created_at: created_at(),
  guild_id: guild_id(),
  inviter: inviter(),
  max_age: max_age(),
  max_uses: max_uses(),
  target_user: target_user(),
  target_user_type: target_user_type(),
  temporary: temporary(),
  uses: uses()
}

target_user()

target_user() :: Nostrum.Struct.User.t() | nil

Partially populated user struct of the target user for this invite.

target_user_type()

target_user_type() :: integer() | nil

The type of user target for this invite.

temporary()

temporary() :: boolean()

Whether this invite only grants temporary membership.

uses()

uses() :: integer()

Number of times this invite has been used.