Nostrum.Struct.Event.MessageReactionAdd (nostrum v0.11.0-dev)

Sent when a user adds a reaction to a message

Types

channel_id()
(since 0.5.0)

@type channel_id() :: Nostrum.Struct.Channel.id()

Channel in which the reaction was added

emoji()
(since 0.5.0)

@type emoji() :: Nostrum.Struct.Emoji.t()

The (partial) emoji used to react

guild_id()
(since 0.5.0)

@type guild_id() :: Nostrum.Struct.Guild.id() | nil

Guild ID in which the reaction was added, if applicable

member()
(since 0.5.0)

@type member() :: Nostrum.Struct.Guild.Member.t() | nil

The member who reacted, if this happened on a guild

message_id()
(since 0.5.0)

@type message_id() :: Nostrum.Struct.Message.id()

Message to which the reaction was added

t()
(since 0.5.0)

@type t() :: %Nostrum.Struct.Event.MessageReactionAdd{
  channel_id: channel_id(),
  emoji: emoji(),
  guild_id: guild_id(),
  member: member(),
  message_id: message_id(),
  user_id: user_id()
}

Event sent when a user adds a reaction to a message

user_id()
(since 0.5.0)

@type user_id() :: Nostrum.Struct.User.id()

ID of the user who added the reaction