Nostrum.Struct.Event.MessageReactionRemove (Nostrum v0.9.0-alpha3)

Sent when a user removes a reaction from a message

Types

channel_id()
(since 0.5.0)

@type channel_id() :: Nostrum.Struct.Channel.id()

ID of the channel in which the reaction was created

emoji()
(since 0.5.0)

@type emoji() :: Nostrum.Struct.Emoji.t() | nil

Partial emoji object that was removed

guild_id()
(since 0.5.0)

@type guild_id() :: Nostrum.Struct.Guild.id() | nil

ID of the guild on which the message lives, if applicable

message_id()
(since 0.5.0)

@type message_id() :: Nostrum.Struct.Message.id()

ID of the message to which the reaction was attached

t()
(since 0.5.0)

@type t() :: %Nostrum.Struct.Event.MessageReactionRemove{
  channel_id: channel_id(),
  emoji: emoji(),
  guild_id: guild_id(),
  message_id: message_id(),
  user_id: user_id()
}

Event sent when a user removes a reaction from a message

user_id()
(since 0.5.0)

@type user_id() :: Nostrum.Struct.User.id()

Author of the reaction