Nostrum.Struct.Event.SpeakingUpdate (Nostrum v0.9.0-alpha3)

Struct representing a Nostrum-generated Speaking Update event

Nostrum will generate this event when the bot starts or stops playing audio.

Types

channel_id()

@type channel_id() :: Nostrum.Struct.Channel.id()

Id of the channel this speaking update is occurring in.

current_url()
(since 0.6.0)

@type current_url() :: String.t() | nil

Current URL being played if a readable format.

guild_id()

@type guild_id() :: Nostrum.Struct.Guild.id()

Id of the guild this speaking update is occurring in.

speaking()

@type speaking() :: boolean()

Boolean representing if bot has started or stopped speaking.

t()

@type t() :: %Nostrum.Struct.Event.SpeakingUpdate{
  channel_id: channel_id(),
  current_url: current_url(),
  guild_id: guild_id(),
  speaking: speaking(),
  timed_out: timed_out()
}

timed_out()
(since 0.5.0)

@type timed_out() :: boolean()

Boolean representing if speaking update was caused by an audio timeout.