Nostrum.Struct.Event.TypingStart (Nostrum v0.10.0)

Sent when a user starts typing in a channel

Types

channel_id()
(since 0.5.0)

@type channel_id() :: Nostrum.Struct.Channel.id()

Channel in which the user started typing

guild_id()
(since 0.5.0)

@type guild_id() :: Nostrum.Struct.Guild.id() | nil

ID of the guild where the user started typing, if applicable

member()
(since 0.5.0)

@type member() :: Nostrum.Struct.Guild.Member.t() | nil

The member who started typing if this happened in a guild

t()
(since 0.5.0)

@type t() :: %Nostrum.Struct.Event.TypingStart{
  channel_id: channel_id(),
  guild_id: guild_id(),
  member: member(),
  timestamp: timestamp(),
  user_id: user_id()
}

Event sent when a user starts typing in a channel

timestamp()
(since 0.5.0)

@type timestamp() :: DateTime.t()

When the user started typing

user_id()
(since 0.5.0)

@type user_id() :: Nostrum.Struct.User.id()

ID of the user who started typing