Nostrum.Struct.Event.VoiceReady (Nostrum v0.4.7)

Struct representing a Nostrum-generated Voice Ready event

Nostrum will generate this event when the bot joins a voice channel and is ready to play audio.

Listening to this event may be used for bots that begin playing audio directly after joining a voice channel as an alternative to waiting until Nostrum.Voice.ready?/1 returns true.

Types

channel_id()

Specs

channel_id() :: Nostrum.Struct.Channel.id()

Id of the channel that voice is ready in.

guild_id()

Specs

guild_id() :: Nostrum.Struct.Guild.id()

Id of the guild that voice is ready in.

t()

Specs

t() :: %Nostrum.Struct.Event.VoiceReady{
  channel_id: channel_id(),
  guild_id: guild_id()
}