Nostrum.Struct.Event.VoiceState (Nostrum v0.10.0)

Represents a user's voice connection status

Types

channel_id()
(since 0.5.0)

@type channel_id() :: Nostrum.Struct.Channel.id()

Channel ID this voice state is for

deaf()
(since 0.5.0)

@type deaf() :: boolean()

Whether this user is deafened by the server

guild_id()
(since 0.5.0)

@type guild_id() :: Nostrum.Struct.Guild.id() | nil

Guild ID this voice state is for, if applicable

member()
(since 0.5.0)

@type member() :: Nostrum.Struct.Guild.Member.t() | nil

Guild member this voice state is for, if applicable

mute()
(since 0.5.0)

@type mute() :: boolean()

Whether this user is muteened by the server

request_to_speak_timestamp()
(since 0.5.0)

@type request_to_speak_timestamp() :: DateTime.t() | nil

Time at which the user requested to speak, if applicable

self_deaf()
(since 0.5.0)

@type self_deaf() :: boolean()

Whether this user is locally deafened

self_mute()
(since 0.5.0)

@type self_mute() :: boolean()

Whether this user is locally muted

self_stream()
(since 0.5.0)

@type self_stream() :: boolean()

Whether the user is streaming using "Go Live"

self_video()
(since 0.5.0)

@type self_video() :: boolean()

Whether this user's camera is enabled

session_id()
(since 0.5.0)

@type session_id() :: String.t()

Session ID for this voice state

suppress()
(since 0.5.0)

@type suppress() :: boolean()

Whether this user is muted by the current user

t()
(since 0.5.0)

@type t() :: %Nostrum.Struct.Event.VoiceState{
  channel_id: channel_id(),
  deaf: deaf(),
  guild_id: guild_id(),
  member: member(),
  mute: mute(),
  request_to_speak_timestamp: request_to_speak_timestamp(),
  self_deaf: self_deaf(),
  self_mute: self_mute(),
  self_stream: self_stream(),
  self_video: self_video(),
  session_id: session_id(),
  suppress: suppress(),
  user_id: user_id()
}

Event sent when a user's voice status is updated

user_id()
(since 0.5.0)

@type user_id() :: Nostrum.Struct.User.id()

User this voice state is for