Elixir v0.3.0 Nostrum.Struct.Guild.Channel

Struct representing a Discord guild channel.

Types

bitrate()

bitrate() :: integer() | nil

The bitrate of the voice channel, nil if text channel

guild_id()

guild_id() :: integer()

The id of the channel’s guild

id()

id() :: integer()

The channel’s id

last_message_id()

last_message_id() :: integer() | nil

Id of the last message sent, nil if voice channel

name()

name() :: String.t()

The name of the channel

permission_overwrites()

permission_overwrites() :: [Nostrum.Struct.Overwrite.t()]

The list of overwrites

position()

position() :: integer()

The ordered position of the channel

t()

t() :: %Nostrum.Struct.Guild.Channel{
  bitrate: bitrate(),
  guild_id: guild_id(),
  id: id(),
  last_message_id: last_message_id(),
  name: name(),
  permission_overwrites: permission_overwrites(),
  position: position(),
  topic: topic(),
  type: type(),
  user_limit: user_limit()
}

topic()

topic() :: String.t() | nil

Current channel topic, nil if voice channel

type()

type() :: Intger

The type of the channel

user_limit()

user_limit() :: integer() | nil

The user limit of the voice channel, nil if text channel

Functions

p_encode()