Nostrum.Struct.Guild.Integration (Nostrum v0.9.0-alpha3)

Struct representing a Discord guild integration.

The struct defined here only has the fields provided for Discord Bot integrations available. If you use Nostrum in a non-bot application, feel free to open an issue to add it.

References

https://discord.com/developers/docs/resources/guild#integration-object

Types

account()
(since 0.5.1)

@type account() :: Nostrum.Struct.Guild.Integration.Account.t()

The integration account.

application()
(since 0.5.1)

@type application() :: Nostrum.Struct.Guild.Integration.Application.t() | nil

The bot/OAuth2 application for discord integrations

enabled()
(since 0.5.0)

@type enabled() :: boolean()

Whether this integration is enabled

guild_id()
(since 0.5.1)

@type guild_id() :: Nostrum.Struct.Guild.id() | nil

The id of the guild this integration is for.

Only included when the Integration is sent via the gateway.

id()
(since 0.5.0)

@type id() :: Nostrum.Snowflake.t()

Snowflake ID of the integration

name()
(since 0.5.0)

@type name() :: String.t()

Name of the integration

t()
(since 0.5.0)

@type t() :: %Nostrum.Struct.Guild.Integration{
  account: account(),
  application: application(),
  enabled: enabled(),
  guild_id: guild_id(),
  id: id(),
  name: name(),
  type: type()
}

Represents a Guild integration

type()
(since 0.5.0)

@type type() :: String.t()

Integration type (Twitch, YouTube or Discord)