Nostrum v0.4.2 Nostrum.Struct.Message

Struct representing a Discord message.

Types

activity()

activity() :: Nostrum.Struct.Message.Activity.t() | nil

The activity of the message. Sent with Rich Presence-related chat embeds.

application()

application() :: Nostrum.Struct.Message.Application.t() | nil

The application of the message. Sent with Rich Presence-related chat embeds.

attachments()

attachments() :: [Nostrum.Struct.Message.Attachment.t()]

List of attached files in the message

author()

author() :: Nostrum.Struct.User.t()

The user struct of the author

channel_id()

channel_id() :: Nostrum.Snowflake.t()

The id of the channel

content()

content() :: String.t()

The content of the message

edited_timestamp()

edited_timestamp() :: String.t() | nil

When the message was edited

embeds()

embeds() :: [Nostrum.Struct.Embed.t()]

List of embedded content in the message

guild_id()

guild_id() :: Nostrum.Snowflake.t() | nil

The id of the guild

id()

id() :: Nostrum.Snowflake.t()

The id of the message

member()

member() :: Nostrum.Struct.Guild.Member.t() | nil

Partial Guild Member object received with the Message Create event if message came from a guild channel.

mention_everyone()

mention_everyone() :: boolean()

Whether this messsage mentions everyone

mention_roles()

mention_roles() :: [Nostrum.Snowflake.t()]

List of roles ids mentioned in the message

mentions()

mentions() :: [Nostrum.Struct.User.t()]

List of users mentioned in the message

nonce()

nonce() :: String.t() | nil

Validates if a message was sent

pinned()

pinned() :: boolean()

Whether this message is pinned

reactions()

reactions() :: [Nostrum.Struct.Message.Reaction.t()] | nil

Reactions to the message.

t()

t() :: %Nostrum.Struct.Message{
  activity: activity(),
  application: application(),
  attachments: attachments(),
  author: author(),
  channel_id: channel_id(),
  content: content(),
  edited_timestamp: edited_timestamp(),
  embeds: embeds(),
  guild_id: guild_id(),
  id: id(),
  member: member(),
  mention_everyone: mention_everyone(),
  mention_roles: mention_roles(),
  mentions: mentions(),
  nonce: nonce(),
  pinned: pinned(),
  reactions: reactions(),
  timestamp: timestamp(),
  tts: tts(),
  type: type(),
  webhook_id: webhook_id()
}

timestamp()

timestamp() :: String.t()

When the message was sent

tts()

tts() :: boolean()

Whether this was a TTS message

type()

type() :: integer()

Type of message.

webhook_id()

webhook_id() :: Nostrum.Snowflake.t() | nil

If the message is generated by a webhook, this is the webhook's id.