Nostrum.Struct.Message.Poll.MediaObject (Nostrum v0.9.0-alpha3)

A struct representing a media item of a poll (e.g. a question or answer)

Types

emoji()

@type emoji() :: %{id: integer() | nil, name: String.t() | nil}

A partial emoji (only supported for answers).

For a custom emoji, only the id field needs to be sent, for a default emoji, only the name field needs to be sent (with the Unicode emoji).

t()

@type t() :: %Nostrum.Struct.Message.Poll.MediaObject{emoji: emoji(), text: text()}

text()

@type text() :: String.t() | nil

Text of the poll media object, either the question or answer text.