Nostrum.Struct.Message.Sticker (Nostrum v0.5.0)

A Nostrum.Struct.Message.Sticker represents a sticker that can be sent inside a Nostrum.Struct.Message.

More information can be found on the Discord API Sticker Object Documentation.

Types

available()

Specs

available() :: boolean()

Whether this guild sticker can be used, may be false due to loss of Server Boosts

description()

Specs

description() :: String.t() | nil

Description of the sticker

format_type()

Specs

format_type() :: integer()

Discord API Sticker Object Format Type Documentation

1 - PNG

2 - APNG

3 - LOTTIE

guild_id()

Specs

guild_id() :: Nostrum.Struct.Guild.id() | nil

Id of the guild that owns this sticker

id()

Specs

id() :: Nostrum.Snowflake.t()

Id of the sticker

name()

Specs

name() :: String.t() | nil

Name of the sticker

pack_id()

Specs

pack_id() :: Nostrum.Snowflake.t()

Id of the pack the sticker is from

sort_value()

Specs

sort_value() :: integer()

The sticker's sort order within its pack

t()

Specs

t() :: %Nostrum.Struct.Message.Sticker{
  available: available(),
  description: description(),
  format_type: format_type(),
  guild_id: guild_id(),
  id: id(),
  name: name(),
  pack_id: pack_id(),
  sort_value: sort_value(),
  tags: tags(),
  type: type(),
  user: user()
}

tags()

Specs

tags() :: String.t() | nil

Discord name of a unicode emoji representing the sticker's expression. for standard stickers, a comma-separated list of related expressions.

type()

Specs

type() :: integer()

Discord API Sticker Object Type Documentation

1 - STANDARD an official sticker in a pack, part of Nitro or in a removed purchasable pack

2 - GUILD a sticker uploaded to a Boosted guild for the guild's members

user()

Specs

user() :: Nostrum.Struct.User.t()

User that uploaded the guild sticker