Elixir v0.3.0 Nostrum.Struct.Snowflake

Functions that work on Snowflakes.

Types

external_snowflake()

external_snowflake() :: String.t()

The type that represents snowflakes in JSON.

In JSON, Snowflakes are typically represented as strings due to some languages not being able to represent such a large number.

t()

t() :: integer()

The snowflake type.

Snowflakes are 64-bit unsigned integers used to represent discord object ids.

Functions

cast(value)

cast(term()) :: {:ok, t() | nil} | :error

Attempts to convert a term into a snowflake.

cast!(value)

cast!(term()) :: t() | nil | no_return()

Same as cast/1, except it raises an ArgumentException on failure.

dump(snowflake)

dump(t()) :: external_snowflake()

Convert a snowflake into its external representation.

is_snowflake(term) (macro)

Returns true if term is a snowflake; otherwise returns false