Nostrum.Struct.Sticker (Nostrum v0.11.0-dev)

A Nostrum.Struct.Sticker represents a sticker that can be sent inside a Nostrum.Struct.Message.

Types

available()
(since 0.10.0)

@type available() :: boolean()

Whether this guild sticker can be used.

May be false due to loss of Server Boosts

description()
(since 0.10.0)

@type description() :: String.t() | nil

Description of the sticker

format_type()
(since 0.10.0)

@type format_type() :: :png | :apng | :lottie | :gif

Format of the sticker.

This field is used to determine the return URL in cdn_url/1.

guild_id()
(since 0.10.0)

@type guild_id() :: Nostrum.Struct.Guild.id() | nil

ID of the guild that owns this sticker.

nil if the sticker is a built-in (type :standard) sticker.

id()
(since 0.10.0)

@type id() :: Nostrum.Snowflake.t()

ID of the sticker

name()
(since 0.10.0)

@type name() :: String.t() | nil

Name of the sticker

pack_id()
(since 0.10.0)

@type pack_id() :: Nostrum.Snowflake.t()

ID of the pack the sticker is from

sort_value()
(since 0.10.0)

@type sort_value() :: integer() | nil

The sticker's sort order within its pack.

Sometimes provided for stickers with type :standard that are in a pack.

t()
(since 0.10.0)

@type t() :: %Nostrum.Struct.Sticker{
  available: available(),
  description: description(),
  format_type: format_type(),
  guild_id: guild_id(),
  id: id(),
  name: name(),
  pack_id: pack_id(),
  sort_value: sort_value(),
  tags: tags(),
  type: type(),
  user: user()
}

tags()
(since 0.10.0)

@type tags() :: String.t()

Tags used by the Discord client to auto-complete a sticker.

For default sticker packs, this is a comma-separated list. For guild stickers, this is the name of the unicode emoji associated by the sticker creator with the sticker.

This is technically a free-text field so consistency in formatting is not guaranteed.

type()
(since 0.10.0)

@type type() :: :standard | :guild

Whether the sticker is a standard (platform made) sticker or a custom guild sticker.

user()
(since 0.10.0)

@type user() :: Nostrum.Struct.User.t() | nil

User that uploaded the guild sticker.

nil if the sticker is a built-in (type :standard) sticker.

Functions

cdn_url(map)
(since 0.10.0)

@spec cdn_url(t()) :: String.t()

Fetch a CDN URL for the sticker object.

:png and :apng stickers will return a .png URL, :gif will return a .gif URL and :lottie will return a .json URL.

Examples

iex> sticker = %Nostrum.Struct.Sticker{format_type: :gif, id: 112233445566778899}
iex> Nostrum.Struct.Sticker.cdn_url sticker
"https://media.discordapp.net/stickers/112233445566778899.gif"

iex> sticker = %Nostrum.Struct.Sticker{format_type: :apng, id: 998877665544332211}
iex> Nostrum.Struct.Sticker.cdn_url sticker
"https://cdn.discordapp.com/stickers/998877665544332211.png"