Nostrum.Struct.Sticker.Pack (Nostrum v0.9.1)

Represents a platform-curated sticker pack on Discord

Types

banner_asset_id()

@type banner_asset_id() :: Nostrum.Snowflake.t()

Asset ID of the banner for this sticker pack.

cover_sticker_id()

@type cover_sticker_id() :: Nostrum.Struct.Sticker.id()

ID of a sticker contained within the pack that should be the cover.

description()

@type description() :: String.t()

Marketing description of the sticker pack.

id()

@type id() :: Nostrum.Snowflake.t()

ID of the sticker pack.

name()

@type name() :: String.t()

Name of the pack.

sku_id()

@type sku_id() :: Nostrum.Snowflake.t()

SKU ID of the sticker pack.

stickers()

@type stickers() :: [Nostrum.Struct.Sticker.t()]

A list of stickers contained within the pack.

t()

@type t() :: %Nostrum.Struct.Sticker.Pack{
  banner_asset_id: banner_asset_id(),
  cover_sticker_id: cover_sticker_id(),
  description: description(),
  id: id(),
  name: name(),
  sku_id: sku_id(),
  stickers: stickers()
}

Functions

banner_url(map)

@spec banner_url(t()) :: String.t()

Return the banner pack URL for a given sticker pack.

This is a marketing banner provided by Discord for their platform curated sticker packs.

Examples

iex> pack = %Nostrum.Struct.Sticker.Pack{banner_asset_id: 112233445566778899}
iex> Nostrum.Struct.Sticker.Pack.banner_url pack
"https://cdn.discordapp.com/app-assets/710982414301790216/store/112233445566778899.png"