Nostrum.Struct.ThreadMember (Nostrum v0.5.1)

Struct representing a thread member object

Types

flags()

Specs

flags() :: non_neg_integer()

Any user-thread settings

guild_id()

Specs

guild_id() :: Nostrum.Struct.Guild.id() | nil

The id of the guild the thread is in.

Only present within THREAD_MEMBER_UPDATE events

id()

Specs

id() :: Nostrum.Snowflake.t() | nil

The id of the thread, omitted within GUILD_CREATE events

join_timestamp()

Specs

join_timestamp() :: DateTime.t()

The timestamp of when the user last joined the thread

t()

Specs

t() :: %Nostrum.Struct.ThreadMember{
  flags: flags(),
  guild_id: guild_id(),
  id: id(),
  join_timestamp: join_timestamp(),
  user_id: user_id()
}

user_id()

Specs

user_id() :: Nostrum.Struct.User.id() | nil

The id of the user, omitted within GUILD_CREATE events