Elixir v0.2.1 Nostrum.TaskedConsumer behaviour

ConsumerSupervisor for gateway event handling.

Consumer vs TaskedConsumer

Unlike Nostrum.Consumer, this module will start up a new task to handle each event, allowing for longer running handles that don’t block the entire consumer.

There’s a limit on the amount of concurrently running tasks. This module acts as a pool of tasks to prevent any overloading while still benefiting from

Event specification

See the documentation in Nostrum.Consumer for more information about events.

Example

An example consumer can be found here.

Functions

start_link(mod)

Callbacks

handle_event(arg0)

handle_event(Nostrum.Consumer.event) :: any