Elixir v0.1.0 Mixcord.Api

Interface for Discord’s rest API.

By default all methods in this module are ran synchronously. If you wish to have async rest operations I recommend you execute these functions inside of a task.

Examples

# Async Task
t = Task.async fn ->
  Mixcord.Api.get_channel_messages(12345678912345, :infinity, {})
end
messages = Task.await t

# A lot of times we don't care about the return value of the function
Task.start fn ->
  messages = ["in", "the", "end", "it", "doesn't", "even", "matter"]
  Enum.each messages, &Mixcord.Api.create_message!(12345678912345, &1)
end

A note about Strings and Ints

Currently, responses from the REST api will have id fields as string. Everything received from the WS connection will have id fields as int.

If you’re processing a response from the API and trying to access something in the cache based off of an id in the response, you will need to conver it to an int using String.to_integer/1. I’m open to suggestions for how this should be handled going forward.

Example

messages = Mixcord.Api.get_pinned_messages!(12345678912345)

authors =
  Enum.map messages, fn msg ->
    author_id = String.to_integer(msg.author.id)
    Mixcord.Cache.User.get!(id: author_id)
  end

Types

error()

error() :: {:error, Mixcord.Error.ApiError.t}

Represents a failed response from the API.

This occurs when hackney or HTTPoison fail, or when the API doesn’t respond with 200 or 204.

limit()

limit() :: Integer.t | :infinity

Represents a limit used to retrieve messages.

Integer number of messages, or :infinity to retrieve all messages.

locator()

locator ::
  {:before, Integer.t} |
  {:after, Integer.t} |
  {:around, Integer.t} |
  {}

Represents a tuple used to locate messages.

The first element of the tuple is an atom. The second element will be a message_id as an integer. The tuple can also be empty to search from the most recent message in the channel

status()

status() :: :dnd | :idle | :online | :invisible

Represents different statuses the bot can have.

:dnd - Red circle.

:idle - Yellow circle.

:online - Green circle.

invisible - The bot will appear offline.

Functions

accept_invite(invite_code)

add_member(guild_id, user_id, options)

add_pinned_message(channel_id, message_id)

add_pinned_message(Integer.t, Integer.t) :: error | {:ok}

Pins a message.

Pins the message specified by message_id in the channel specified by channel_id.

Returns {:ok} if successful. error otherwise.

add_pinned_message!(channel_id, message_id)

add_pinned_message!(Integer.t, Integer.t) :: no_return | {:ok}

Pins a message.

Pins the message specified by message_id in the channel specified by channel_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

bangify(to_bang)

batch_modify_guild_roles(guild_id, options)

begin_guild_prune(guild_id, options)

bulk_delete_messages(channel_id, messages)

bulk_delete_messages(Integer.t, [Mixcord.Struct.Message.id]) ::
  error |
  {:ok}

bulk_delete_messages(Integer.t, [Mixcord.Struct.Message.id]) ::
  no_return |
  {:ok}

Deletes multiple messages from a channel.

messages is a list of Mixcord.Struct.Message.id that you wish to delete.

bulk_delete_messages!(channel_id, messages)

Deletes multiple messages from a channel.

See bulk_delete_messages/2 for more info.

Raises Mixcord.Error.ApiError if error occurs while making the rest call.

create_channel(guild_id, options)

create_channel_invite(channel_id, options \\ %{})

create_channel_invite!(channel_id, options \\ %{})

create_dm()

create_guild_ban(guild_id, user_id, options)

create_guild_integrations(guild_id, options)

create_guild_roles(guild_id)

create_message(channel_id, content, tts \\ false)

create_message(Integer.t, String.t | [content: String.t, embed: Mixcord.Struct.Embed] | [content: String.t, file: String.t], boolean) ::
  error |
  {:ok, Mixcord.Struct.Message.t}

Send a message to a channel.

Send content to the channel identified with channel_id. Content is a binary containing the message you want to send. For embeds or file uploads, content should be a keyword list.

tts is an optional parameter that dictates whether the message should be played over text to speech.

Example

Mixcord.Api.create_message(1111111111111, [content: "my os rules", file: ~S"C:\i\use\windows"])

Returns {:ok, Mixcord.Struct.Message} if successful. error otherwise.

create_message!(channel_id, content, tts \\ false)

create_message!(Integer.t, String.t, boolean) ::
  no_return |
  Mixcord.Struct.Message.t

Send a message to a channel.

Send content to the channel identified with channel_id. tts is an optional parameter that dictates whether the message should be played over text to speech.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns Mixcord.Struct.Message if successful.

create_reaction(channel_id, message_id, emoji)

create_reaction(Integer.t, Integer.t, String.t | Mixcord.Struct.Emoji.custom_emoji) ::
  error |
  {:ok}

create_reaction(Integer.t, Integer.t, String.t | Mixcord.Struct.Emoji.custom_emoji) ::
  error |
  {:ok}

Create a rection for a message.

Creates a reaction using an emoji for the message specified by message_id and channel_id. emoji can be a Mixcord.Struct.Emoji.custom_emoji.t or a base 16 unicode emoji string.

Example

Mixcord.Api.create_reaction(123123123123, 321321321321, "\xF0\x9F\x98\x81")

Returns {:ok} if successful, {:error, reason} otherwise.

delete_all_reactions(channel_id, message_id)

delete_all_reactions(Integer.t, Integer.t) :: error | {:ok}

Deletes all reactions from a message.

Reaction to delete is specified by channel_id, message_id, and emoji.

Returns {:ok} if successful, {:error, reason} otherwise.

delete_channel(channel_id)

delete_channel(Integer.t) ::
  error |
  {:ok, Mixcord.Struct.Channel.t}

Delete a channel.

Channel to delete is specified by channel_id.

delete_channel!(channel_id)

delete_channel!(Integer.t) ::
  no_return |
  Mixcord.Struct.Channel.t

Delete a channel.

Raises Mixcord.Error.ApiError if error occurs while making the rest call.

delete_channel_permissions(channel_id, overwrite_id)

delete_channel_permissions!(channel_id, overwrite_id)

delete_guild(guild_id)

delete_guild_integrations(guild_id, integration_id)

delete_guild_role(guild_id, role_id)

delete_invite(invite_code)

delete_message(channel_id, message_id)

delete_message(Integer.t, Integer.t) :: error | {:ok}

Delete a message.

Delete a message specified by channel_id and message_id.

Returns {:ok} if successful. error otherwise.

delete_message!(channel_id, message_id)

delete_message!(String.t, Integer.t) :: no_return | {:ok}

Delete a message.

Delete a message specified by channel_id and message_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

delete_own_reaction(channel_id, message_id, emoji)

Deletes a rection made by the user.

Reaction to delete is specified by channel_id, message_id, and emoji.

Returns {:ok} if successful, {:error, reason} otherwise.

delete_pinned_message(channel_id, message_id)

delete_pinned_message(Integer.t, Integer.t) :: error | {:ok}

Unpins a message.

Unpins the message specified by message_id in the channel specified by channel_id.

Returns {:ok} if successful. error otherwise.

delete_pinned_message!(channel_id, message_id)

delete_pinned_message!(Integer.t, Integer.t) ::
  no_return |
  {:ok}

Unpins a message.

Unpins the message specified by message_id in the channel specified by channel_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

delete_reaction(channel_id, message_id, emoji, user_id)

delete_reaction(Integer.t, Integer.t, String.t | Mixcord.Struct.Emoji.custom_emoji, Integer.t) ::
  error |
  {:ok}

Deletes a rection from a message.

Reaction to delete is specified by channel_id, message_id, emoji, and user_id.

Returns {:ok} if successful, {:error, reason} otherwise.

edit_channel(channel_id, options)

edit_channel(Integer.t, name: String.t, position: Integer.t, topic: String.t, bitrate: String.t, user_limit: Integer.t) ::
  error |
  {:ok, Mixcord.Struct.Channel.t}

Edit a channel.

Edits a channel with options

options is a kwl with the following optional keys:

name - New name of the channel.

position - Position of the channel.

topic - Topic of the channel. Text Channels only

bitrate - Bitrate of the voice channel. Voice Channels only

user_limit - User limit of the channel. 0 for no limit. Voice Channels only

edit_channel!(channel_id, options)

edit_channel!(Integer.t, name: String.t, position: Integer.t, topic: String.t, bitrate: String.t, user_limit: Integer.t) ::
  error |
  {:ok, Mixcord.Struct.Channel.t}

Edit a channel.

See edit_channel/2 for parameters.

Raises Mixcord.Error.ApiError if error occurs while making the rest call.

edit_channel_permissions(channel_id, overwrite_id, permission_info)

edit_channel_permissions(Integer.t, Integer.t, allow: Integer.t, deny: Integer.t, type: String.t) ::
  error |
  {:ok}

Edit the permission overwrites for a user or role.

Role or user to overwrite is specified by channel_id and overwrite_id.

permission_info is a kwl with the following required keys:

allow - Bitwise value of allowed permissions.

deny - Bitwise value of denied permissions.

type - member if editing a user, role if editing a role.

edit_channel_permissions!(channel_id, overwrite_id, permission_info)

edit_channel_permissions!(Integer.t, Integer.t, allow: Integer.t, deny: Integer.t, type: String.t) ::
  no_return |
  {:ok}

Edit the permission overwrites for a user or role.

See edit_channel_permissions/2 for more info.

Raises Mixcord.Error.ApiError if error occurs while making the rest call.

edit_guild(guild_id, options)

edit_message(channel_id, message_id, content)

edit_message(Integer.t, Integer.t, String.t) ::
  error |
  {:ok, Mixcord.Struct.Message.t}

Edit a message.

Edit a message with the given content. Message to edit is specified by channel_id and message_id.

Returns the edited {:ok, Mixcord.Struct.Message} if successful. error otherwise.

edit_message!(channel_id, message_id, content)

edit_message!(Integer.t, Integer.t, String.t) ::
  no_return |
  {:ok, Mixcord.Struct.Message.t}

Edit a message.

Edit a message with the given content. Message to edit is specified by channel_id and message_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns the edited Mixcord.Struct.Message if successful.

get_channel(channel_id)

get_channel(Integer.t) ::
  error |
  {:ok, Mixcord.Struct.Channel.t}

Get a channel.

Gets a channel specified by id.

get_channel!(channel_id)

get_channel!(Integer.t) :: no_return | Mixcord.Struct.Channel.t

Get a channel.

Gets a channel specified by id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call.

get_channel_invites(channel_id)

get_channel_invites!(channel_id)

get_channel_message(channel_id, message_id)

get_channel_message(Integer.t, Integer.t) ::
  error |
  {:ok, Mixcord.Struct.Message.t}

Retrieves a message from a channel.

Message to retrieve is specified by message_id and channel_id.

get_channel_message!(channel_id, message_id)

get_channel_message!(Integer.t, Integer.t) ::
  no_return |
  Mixcord.Struct.Message.t

Retrieves a message from a channel.

Raises Mixcord.Error.ApiError if error occurs while making the rest call.

get_channel_messages(channel_id, limit, locator)

get_channel_messages(Integer.t, limit, locator) ::
  error |
  {:ok, [Mixcord.Struct.Message.t]}

Retrieve messages from a channel.

Retrieves limit number of messages from the channel with id channel_id. locator is a tuple indicating what messages you want to retrieve.

Returns {:ok, [Mixcord.Struct.Message]} if successful. error otherwise.

get_channel_messages!(channel_id, limit, locator)

get_channel_messages!(Integer.t, limit, locator) ::
  no_return |
  [Mixcord.Struct.Message.t]

Retrieve messages from a channel.

See get_channel_message/3 for usage.

Raises Mixcord.Error.ApiError if error occurs while making the rest call.

get_channel_messages_call(channel_id, limit, locator)

get_current_user()

get_current_users_guilds()

get_guild(guild_id)

get_guild(Integer.t) :: error | {:ok, Mixcord.Struct.Guild.t}

Gets a guild using the REST api

Retrieves a guild with specified guild_id.

Returns {:ok, Mixcord.Struct.Guild.t} if successful, error otherwise.

get_guild!(guild_id)

get_guild!(Integer.t) :: no_return | Mixcord.Struct.Guild.t

Gets a guild using the REST api

Retrieves a guild with specified guild_id.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns Mixcord.Struct.Guild.t if successful.

get_guild_bans(guild_id)

get_guild_embed(guild_id)

get_guild_integrations(guild_id)

get_guild_invites(guild_id)

get_guild_prune(guild_id, options)

get_guild_roles(guild_id)

get_invite(invite_code)

get_member(guild_id, user_id)

get_pinned_messages(channel_id)

get_pinned_messages(Integer.t) ::
  error |
  {:ok, [Mixcord.Struct.Message.t]}

Gets all pinned messages.

Retrieves all pinned messages for the channel specified by channel_id.

Returns {:ok, [Mixcord.Struct.Message.t]} if successful. error otherwise.

get_pinned_messages!(channel_id)

get_pinned_messages!(Integer.t) ::
  no_return |
  [Mixcord.Struct.Message.t]

Gets all pinned messages.

Retrieves all pinned messages for the channel specified by channel_id.

Returns [Mixcord.Struct.Message.t] if successful. error otherwise.

get_reactions(channel_id, message_id, emoji)

get_reactions(Integer.t, Integer.t, String.t | Mixcord.Struct.Emoji.custom_emoji) ::
  error |
  {:ok, [Mixcord.Struct.User]}

Gets all users who reacted with an emoji.

Retrieves a list of users who have reacted with an emoji.

Returns {:ok, [Mixcord.Struct.User]} if successful, {:error, reason} otherwise.

get_token()

get_token() :: String.t

Returns the token of the bot.

get_user(user_id)

get_user_connections()

get_user_dms()

get_voice_region(guild_id)

guild_members(guild_id, options)

leave_guild(guild_id)

list_voice_regions()

modify_channel_position(guild_id, options)

modify_current_user()

modify_guild_embed(guild_id)

modify_guild_integrations(guild_id, integration_id, options)

modify_guild_roles(guild_id, role_id, options)

modify_member(guild_id, user_id, options)

remove_guild_ban(guild_id, user_id)

remove_member(guild_id, user_id)

request(method, route, body \\ "", options \\ [])

request_multipart(method, route, body \\ "", options \\ [])

start_typing(channel_id)

start_typing(Integer.t) :: error | {:ok}

Triggers the typing indicator.

Triggers the typing indicator in the channel specified by channel_id. The typing indicator lasts for about 8 seconds and then automatically stops.

Returns {:ok} if successful. error otherwise.

start_typing!(channel_id)

start_typing!(Integer.t) :: no_return | {:ok}

Triggers the typing indicator.

Triggers the typing indicator in the channel specified by channel_id. The typing indicator lasts for about 8 seconds and then automatically stops.

Raises Mixcord.Error.ApiError if error occurs while making the rest call. Returns {:ok} if successful.

sync_guild_integrations(guild_id, integration_id)

update_status(status, game)

update_status(status, String.t) :: no_return

Updates the status of the bot for all shards.

For more information see Mixcord.Api.update_status/3

update_status(pid, status, game)

update_status(Pid.t, status, String.t) :: no_return

Updates the status of the bot for a certain shard.

pid is the pid of the shard whose status you want to update. To update the status for all shards see Mixcord.Api.update_status/2 status is an atom that describes the status of the bot. See Mixcord.Api.status.t for available options. game is the text that will display ‘playing’ status of the game. This is the text below the bot’s name in the sidebar. Empty string will clear.