Elixir v0.1.0 Mixcord.Shard.Dispatch.Consumer behaviour

Consumer process for gateway event handling.

Consuming Gateway Events

To handle events, Mixcord uses a GenStage implementation. GenStage is “new” with Elixir version 1.4, expanding on the old functionality of GenEvent.

Mixcord defines the producer in the GenStage design. To consume the events you must create at least one consumer process. For an example of this behaviour please see here.

It is generally recommended that you spawn a consumer per core. To find this number you can use System.schedulers_online/0.

Types

build_ban_remove()

build_ban_remove() :: {:GUILD_BAN_REMOVE, {Mixcord.Struct.User.t}, ws_state}

channel_create()

channel_create() :: {:CHANNEL_CREATE, {Mixcord.Struct.Channel.t}, ws_state}

channel_delete()

channel_delete() :: {:CHANNEL_DELETE, {Mixcord.Struct.Channel.t}, ws_state}

channel_pins_ack()

channel_pins_ack() :: {:CHANNEL_PINS_ACK, {Map.t}, ws_state}

channel_pins_update()

channel_pins_update() :: {:CHANNEL_PINS_UPDATE, {Map.t}, ws_state}

channel_update()

channel_update() :: {:CHANNEL_UPDATE, {old_channel :: Mixcord.Struct.Channel.t, new_channel :: Mixcord.Struct.Channel.t}, ws_state}

event()

event ::
  channel_create |
  channel_delete |
  channel_update |
  channel_pins_ack |
  channel_pins_update |
  guild_ban_add |
  build_ban_remove |
  guild_create |
  guild_update |
  guild_delete |
  guild_emojis_update |
  guild_integrations_update |
  guild_member_add |
  guild_members_chunk |
  guild_member_remove |
  guild_member_update |
  guild_role_create |
  guild_role_delete |
  guild_role_update |
  message_create |
  message_delete |
  message_delete_bulk |
  message_update |
  presence_update |
  ready |
  resumed |
  typing_start |
  user_settings_update |
  user_update |
  voice_state_update |
  voice_server_update

from()

from() :: {pid, tag :: term}

Tuple describing the client of a call request.

pid is the PID of the caller and tag is a unique term used to identify the call.

guild_ban_add()

guild_ban_add() :: {:GUILD_BAN_ADD, {Mixcord.Struct.User.t}, ws_state}

guild_create()

guild_create() :: {:GUILD_CREATE, {new_guild :: Mixcord.Struct.Guild.t}, ws_state}

guild_delete()

guild_delete() :: {:GUILD_DELETE, {old_guild :: Mixcord.Struct.Guild.t}, ws_state}

guild_emojis_update()

guild_emojis_update() :: {:GUILD_EMOJIS_UPDATE, {old_emojis :: [Mixcord.Struct.Emoji.t], new_emojis :: [Mixcord.Struct.Emoji.t]}, ws_state}

guild_integrations_update()

guild_integrations_update() :: {:GUILD_INTEGERATIONS_UPDATE, {Map.t}, ws_state}

guild_member_add()

guild_member_add() :: {:GUILD_MEMBER_ADD, {new_member :: Mixcord.Struct.Member.t}, ws_state}

guild_member_remove()

guild_member_remove() :: {:GUILD_MEMBER_REMOVE, {old_member :: Mixcord.Struct.Member.t}, ws_state}

guild_member_update()

guild_member_update() :: {:GUILD_MEMBER_UPDATE, {old_member :: Mixcord.Struct.Member.t, new_member :: Mixcord.Struct.Member.t}, ws_state}

guild_members_chunk()

guild_members_chunk() :: {:GUILD_MEMBERS_CHUNK, {Map.t}, ws_state}

guild_role_create()

guild_role_create() :: {:GUILD_ROLE_CREATE, {new_role :: Mixcord.Struct.Role.t}, ws_state}

guild_role_delete()

guild_role_delete() :: {:GUILD_ROLE_DELETE, {old_role :: Mixcord.Struct.Role.t}, ws_state}

guild_role_update()

guild_role_update() :: {:GUILD_ROLE_UPDATE, {old_role :: Mixcord.Struct.Role.t, new_role :: Mixcord.Struct.Role.t}, ws_state}

guild_update()

guild_update() :: {:GUILD_CREATE, {old_guild :: Mixcord.Struct.Guild.t, new_guild :: Mixcord.Struct.Guild.t}, ws_state}

message_create()

message_create() :: {:MESSAGE_CREATE, {message :: Mixcord.Struct.Message.t}, ws_state}

message_delete()

message_delete() :: {:MESSAGE_DELETE, {message :: Mixcord.Struct.Message.t}, ws_state}

message_delete_bulk()

message_delete_bulk() :: {:MESSAGE_DELETE_BULK, {updated_messages :: [Mixcord.Struct.Message.t]}, ws_state}

message_update()

message_update() :: {:MESSAGE_UPDATE, {updated_message :: Mixcord.Struct.Message.t}, ws_state}

presence_update()

presence_update() :: {:PRESENCE_UPDATE, {Map.t}, ws_state}

ready()

ready() :: {:READY, {Map.t}, ws_state}

resumed()

resumed() :: {:RESUMED, {Map.t}, ws_state}

state()

state() :: Map.t

The state of the consumer process.

typing_start()

typing_start() :: {:TYPING_START, {Map.t}, ws_state}

user_settings_update()

user_settings_update() :: no_return

user_update()

user_update() :: {:USER_UPDATE, {old_user :: Mixcord.Struct.User.t, new_user :: Mixcord.Struct.User.t}, ws_state}

voice_server_update()

voice_server_update() :: {:VOICE_SERVER_UPDATE, {Map.t}, ws_state}

voice_state_update()

voice_state_update() :: {:VOICE_STATE_UPDATE, {Map.t}, ws_state}

ws_state()

ws_state() :: Map.t

The state of the websocket connection for the shard the event occured on.

Functions

handle_events(events, from, state)

Invoked on :producer_consumer and :consumer stages to handle events.

Must always be explicitly implemented by such types.

Return values are the same as c:handle_cast/2.

Callback implementation for GenStage.handle_events/3.

init(state)

Invoked when the server is started.

start_link/3 (or start/3) will block until it returns. args is the argument term (second argument) passed to start_link/3.

In case of successful start, this callback must return a tuple where the first element is the stage type, which is either a :producer, :consumer or :producer_consumer if it is taking both roles.

For example:

def init(args) do
  {:producer, some_state}
end

The returned tuple may also contain 3 or 4 elements. The third element may be the :hibernate atom or a set of options defined below.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling terminate/2.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling terminate/2.

Options

This callback may return options. Some options are specific to the stage type while others are shared across all types.

:producer options

:demand - when :forward, the demand is always forwarded to the handle_demand callback. When :accumulate, demand is accumulated until its mode is set to :forward via demand/2. This is useful as a synchronization mechanism, where the demand is accumulated until all consumers are subscribed. Defaults to :forward.

:producer and :producer_consumer options

:buffer_size - the size of the buffer to store events without demand. Check the “Buffer events” section on the module documentation (defaults to 10000 for :producer, :infinity for :producer_consumer)

:buffer_keep - returns if the :first or :last (default) entries should be kept on the buffer in case we exceed the buffer size

:dispatcher - the dispatcher responsible for handling demands. Defaults to GenStage.DemandDispatch. May be either an atom or a tuple with the dispatcher and the dispatcher options

:consumer and :producer_consumer options

:subscribe_to - a list of producers to subscribe to. Each element represents the producer or a tuple with the producer and the subscription options (as defined in sync_subscribe/2)

Dispatcher

When using a :producer or :producer_consumer, the dispatcher may be configured on init as follows:

{:producer, state, dispatcher: GenStage.BroadcastDispatcher}

Some dispatchers may require options to be given on initialization, those can be done with a tuple:

{:producer, state, dispatcher: {GenStage.PartitionDispatcher, partitions: 0..3}}

Callback implementation for GenStage.init/1.

start_link(mod)

Callbacks

handle_event(event, state)

handle_event(event, state) :: {:ok, Map.t}

Callback used to handle events.

Event is the event name as an atom, and ws_state is the current state of the websocket that the event was received on. For more information on this please see Mixcord.Shard.Payload.state_map.t.

from is the process information of the producer from which the demand was received. state is the internal state of your consumer.