Nostrum.Api.Ratelimiter (Nostrum v0.4.5)

Ratelimit implimentation specific to Discord's API. Only to be used when starting in a rest-only manner.

Types

on_start()

Specs

on_start() ::
  {:ok, pid()} | :ignore | {:error, {:already_started, pid()} | term()}

Return values of start functions.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

empty_buckets()

Specs

empty_buckets() :: true

Empties all buckets, voiding any saved ratelimit values.

get_endpoint(route, method)

Specs

get_endpoint(String.t(), atom()) :: String.t()

Retrieves a proper ratelimit endpoint from a given route and url.

init(list)

Callback implementation for GenServer.init/1.

start_link(list)

Specs

start_link([]) :: on_start()

Starts the ratelimiter.