Elixir v0.3.0 Nostrum.Cache.ChannelCache

Cache for channels.

The ETS table name associated with the Channel Cache is :channels. Besides the methods provided below you can call any other ETS methods on the table.

Example

info = :ets.info(:channels)
[..., heir: :none, name: :channels, size: 1, ...]
size = info[:size]
1

Functions

get(id)

get(Nostrum.Struct.Channel.id() | Nostrum.Struct.Message.t()) ::
  {:error, atom()} | {:ok, Nostrum.Struct.Channel.t()}

Retrieves a channel from the cache.

Internally, the ChannelCache process only stores Nostrum.Struct.Channel.dm_channel/0 references. To get channel information, a call is made to a Nostrum.Cache.GuildCache.

If successful, returns {:ok, channel}. Otherwise, returns {:error, reason}

Example

case Nostrum.Cache.ChannelCache.get(133333333337) do
  {:ok, channel} ->
    "We found " <> channel.name
  {:error, _reason} ->
    "Donde esta"
end

get!(id)

get!(Nostrum.Struct.Channel.id() | Nostrum.Struct.Message.t()) ::
  no_return() | Nostrum.Struct.Channel.t()

Same as get/1, but raises Nostrum.Error.CacheError in case of a failure.