Nostrum.Cache.ChannelCache.ETS (Nostrum v0.9.0-alpha3)

An ETS-based cache for channels outside of guilds.

The supervisor defined by this module will set up the ETS table associated with it.

If you need to get the table reference for the table used by this module, please use the table/0 function.

Note that users should not call the functions not related to this specific implementation of the cache directly. Instead, call the functions of Nostrum.Cache.ChannelCache directly, which will dispatch to the configured cache.

Functions

child_spec(init_arg)
(since 0.5.0)

Returns a specification to start this module under a supervisor.

See Supervisor.

create(channel)
(since 0.5.0)

@spec create(map()) :: Nostrum.Struct.Channel.t()

Converts and creates the given map as a channel in the cache.

delete(id)
(since 0.5.0)

@spec delete(Nostrum.Struct.Channel.id()) :: :noop | Nostrum.Struct.Channel.t()

Delete the channel from the cache by ID.

init(init_arg)
(since 0.5.0)

Set up the cache's ETS table.

query_handle()
(since 0.8.0)

@spec query_handle() :: :qlc.query_handle()

Retrieve a query handle for usage with QLC.

start_link(init_arg)
(since 0.5.0)

Start the supervisor.

table()
(since 0.8.0)

@spec table() :: :ets.table()

Retrieve the ETS table reference used for the cache.

update(channel)
(since 0.5.0)

@spec update(Nostrum.Struct.Channel.t()) ::
  {Nostrum.Struct.Channel.t() | nil, Nostrum.Struct.Channel.t()}

Update the given channel in the cache.