Elixir v0.3.0 Nostrum.Cache.Guild.GuildServer

Module for interacting with Guild Servers.

Types

search_criteria()

search_criteria() ::
  [{:id, integer()}]
  | [{:channel_id, integer()}]
  | [{:message, Nostrum.Struct.Message.t()}]

Represents different ways to search for a guild.

Examples

Search by Nostrum.Struct.Guild.id/0

  GuildServer.get(id: 666666666666666)

Search by t:Nostrum.Struct.Guild.TextChannel.id/0

  GuildServer.get(channel_id: 314159265358979)

Search via Nostrum.Struct.Message

  msg = Nostrum.Api.create_message!(1618033988731, "Drake confirmed British")
  GuildServer.get(msg)

search_criteria_with_key()

search_criteria_with_key() ::
  [id: integer(), key: key :: atom()]
  | [channel_id: integer(), key: key :: atom()]
  | [message: Nostrum.Struct.Message.t(), key: key :: atom()]

Represents different ways to get a value from a guild.

Same as search_criteria/0, but takes an additional key parameter. Keys can be found in Nostrum.Struct.Guild

search_criteria_with_transform()

search_criteria_with_transform() ::
  [id: integer(), transform: transform()]
  | [channel_id: integer(), transform: transform()]
  | [message: Nostrum.Struct.Message.t(), transform: transform()]

Represents different ways to transform a guild.

Same as search_criteria/0, but takes an additional transform parameter. See transform/1 for an example.

transform()

transform() :: (Nostrum.Struct.Guild.t() -> term())

Transform for a guild.

The function is passed a Nostrum.Struct.Guild.t struct.

Functions

all()

all() :: Enumerable.t()

Retrieves a stream of all guilds.

This will introduce a very large binary into the caller process. Consider running this method inside its own task.

If you just need to get a specific key, consider using get_value_all/1

Example

Nostrum.Cache.Guild.GuildServer.all
|> Enum.take_every(3)
|> D3L3T3_GU1LD

cast(id, request)

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get(search_criteria)

get(search_criteria()) ::
  {:error, reason :: atom()} | {:ok, Nostrum.Struct.Guild.t()}

Retrieves a guild from the cache.

Parameters

search_criteria - The criteria used to search. See search_criteria/0 for more info.

Example

case Nostrum.Cache.Guild.GuildServer.get(id: 1234567891234789) do
  {:ok, guild} -> "Guild has #{length(guild.members)} members"
  {:error, _reason} -> "Guild is MIA"
end

get!(search_criteria)

get!(search_criteria()) :: no_return() | {Nostrum.Struct.Guild.t()}

Retrieves a guild from the cache.

See Nostrum.Cache.Guild.GuildServer.get/1 for usage.

Raises Nostrum.Error.CacheError if unable to find the guild.

get_value(list)

get_value(search_criteria_with_key()) ::
  {:error, reason :: atom()} | {:ok, term()}

Retrieves a value from a guild in the cache.

This method should be preferred over the get/1 method in the event that you need only one kv pair from the guild.

Parameters

search_criteria_with_key - The criteria used to search. See search_criteria_with_key/0 for more info.

Example

case Nostrum.Cache.Guild.Guildserver.get_value(id: 69696969696969, key: :name) do
  {:ok, sick_guild_name} -> "#{sick_guild_name} is sick"
  {:error, reason} -> "Whatchu' searchin' for fam?"
end

get_value!(search_critera_with_key)

get_value!(search_criteria_with_key()) :: no_return() | term()

Retrieves a value from a guild in the cache.

See Nostrum.Cache.Guild.GuildServer.get_value!/2 for usage.

Raises Nostrum.Error.CacheError if unable to find the guild or key.

get_value_from_all(key)

get_value_from_all(key :: atom()) :: Enumerable.t()

Retrieves a stream of values from all guilds.

Parameter

key - Value to get from all guilds.

Example

names =
Nostrum.Cache.GuildServer.get_value_from_all(:name)
|> Enum.filter(&is_weeb?)
|> D3L3T3_GU1LD5

index_guild(guild)

member_chunk(guild_id, user)

transform(list)

transform(search_criteria_with_transform()) :: {:ok, term()}

Retrieves a transformed guild from the cache.

Returns the result of applying a given fun to the guild specified by the search criteria.

This transform does not modify the cache and is performed entirely server side. This means that if you somehow cause an error to be thrown, the receiving guild process will die. The benefit is not “copying” guild binaries between processes which can be very big.

Parameters

search_criteria_with_transform - The criteria used to search. See search_criteria_with_transform/0 for more info.

Example

tf = fn guild -> Enum.any?(guild.members, &(Map.get(&1, :nick) == "The year is 20XX")) end
{:ok, plays_fox?} = Nostrum.Cache.Guild.GuildServer.transform(id: 123, transform: tf)

transform_all(transform)

transform_all(transform()) :: Enumerable.t()

Retrieves a stream of transformed guild values from the cache.

Returns a stream of as the result of calling transform on every guild in the cache.

See transform/1 for usage.