Nostrum.Cache.GuildCache behaviour (Nostrum v0.5.0)

Cache behaviour & dispatcher for guilds.

You can call the functions provided by this module independent of which cache is configured, and it will dispatch to the configured cache implementation.

By default, Elixir.Nostrum.Cache.GuildCache.ETS will be used for caching guilds. You can override this in the :caches option of the :nostrum application by setting the :guilds field to a different module implementing the Nostrum.Cache.GuildCache behaviour. Any module below Nostrum.Cache.GuildCache can be used as a cache.

Writing your own guild cache

As with the other caches, the guild cache API consists of two parts:

The functions that the user calls, such as all/0 or select_by/2.

The functions that nostrum calls, such as create/1 or update/1. These do not create any objects in the Discord API, they are purely created to update the cached data from data that Discord sends us. If you want to create objects on Discord, use the functions exposed by Nostrum.Api instead.

You need to implement both of them for nostrum to work with your custom cache. You also need to implement Supervisor callbacks, which will start your cache as a child under Nostrum.Cache.CacheSupervisor: As an example, the Nostrum.Cache.GuildCache.ETS implementation uses this to to set up its ETS table it uses for caching. See the callbacks section for every nostrum-related callback you need to implement.

Note that this module also defines a few helper functions, such as get!/1 or select_by!/2, which call the backing cache's regular functions and perform the result unwrapping by themselves.

The "upstream data" wording in this module references the fact that the data that the guild cache (and other caches) retrieves represents the raw data we receive from the upstream connection, no attempt is made by nostrum to sanitize the data before it enters the cache. Caching implementations need to cast the data to the resulting type themselves. A possible future improvement would be moving the data casting into this module before the backing cache implementation is called.

Types

clause()

Specs

clause() ::
  {:id, Nostrum.Struct.Guild.id()}
  | {:channel_id, Nostrum.Struct.Channel.id()}
  | {:message, Nostrum.Struct.Message.t()}

A clause for filtering guilds.

clauses()

Specs

clauses() :: [clause()] | map()

A collection of clause/0s for filtering guilds.

reason()

Specs

reason() :: :id_not_found | :id_not_found_on_guild_lookup

Specifies the reason for why a lookup operation has failed.

selector()

Specs

selector() :: (Nostrum.Struct.Guild.t() -> any())

A selector for looking up entries in the cache.

Callbacks

all()

Specs

all() :: Enum.t()

Retrieves all Nostrum.Struct.Guild from the cache.

channel_create(id, channel)

Specs

channel_create(Nostrum.Struct.Guild.id(), channel :: map()) ::
  Nostrum.Struct.Channel.t()

Create a channel for the guild from upstream data.

Return the adapted Nostrum.Struct.Channel.t/0 structure.

channel_delete(id, id)

Specs

channel_delete(Nostrum.Struct.Guild.id(), Nostrum.Struct.Channel.id()) ::
  Nostrum.Struct.Channel.t() | :noop

Delete the given channel from the guild.

If the channel was cached, return the original channel. Return :noop otherwise.

channel_update(id, channel)

Specs

channel_update(Nostrum.Struct.Guild.id(), channel :: map()) ::
  {old_channel :: Nostrum.Struct.Channel.t(),
   new_channel :: Nostrum.Struct.Channel.t()}

Update the given channel on the given guild from upstream data.

Return the original channel before the update, and the updated channel.

create(t)

Specs

create(Nostrum.Struct.Guild.t()) :: true

Create a guild in the cache.

delete(id)

Specs

delete(Nostrum.Struct.Guild.id()) :: Nostrum.Struct.Guild.t() | nil

Delete a guild from the cache.

Return the old guild if it was cached, or nil otherwise.

emoji_update(id, emojis)

Specs

emoji_update(Nostrum.Struct.Guild.id(), emojis :: [map()]) ::
  {old_emojis :: [Nostrum.Struct.Emoji.t()],
   new_emojis :: [Nostrum.Struct.Emoji.t()]}

Update the emoji list of the given guild from upstream data.

Discord sends us the complete emoji list on an update, which is passed here.

Return the old list of emojis before the update, and the updated list of emojis.

get(id)

Specs

get(Nostrum.Struct.Guild.id()) ::
  {:ok, Nostrum.Struct.Guild.t()} | {:error, reason()}

Retrives a single Nostrum.Struct.Guild from the cache via its id.

Returns {:error, reason} if no result was found.

Examples

iex> Nostrum.Cache.GuildCache.get(0)
{:ok, %Nostrum.Struct.Guild{id: 0}}

iex> Nostrum.Cache.GuildCache.get(10)
{:error, :id_not_found_on_guild_lookup}

get_by(clauses)

Specs

get_by(clauses()) :: {:ok, Nostrum.Struct.Guild.t()} | {:error, reason()}

Retrives a single Nostrum.Struct.Guild where it matches the clauses.

Returns {:error, reason} if no result was found.

iex> Nostrum.Cache.GuildCache.get_by(id: 0)
{:ok, %Nostrum.Struct.Guild{id: 0}}

iex> Nostrum.Cache.GuildCache.get_by(%{id: 0})
{:ok, %Nostrum.Struct.Guild{id: 0}}

iex> Nostrum.Cache.GuildCache.get_by(id: 10)
{:error, :id_not_found_on_guild_lookup}

member_add(id, member)

Specs

member_add(Nostrum.Struct.Guild.id(), member :: map()) ::
  Nostrum.Struct.Guild.Member.t()

Add the member for the given guild from upstream data.

Return the casted member structure.

member_chunk(id, chunk)

Specs

member_chunk(Nostrum.Struct.Guild.id(), chunk :: [member :: map()]) :: true

Bulk create multiple members in the cache from upstream data.

Return value is unused, as we currently do not dispatch a gateway for this.

member_remove(id, member)

Specs

member_remove(Nostrum.Struct.Guild.id(), member :: map()) ::
  {Nostrum.Struct.Guild.id(), old_member :: Nostrum.Struct.Guild.Member.t()}
  | :noop

Remove the given member for the given guild from upstream data.

Return the guild ID and old member if the member was cached. Otherwise, return :noop.

member_update(id, member)

Specs

member_update(Nostrum.Struct.Guild.id(), member :: map()) ::
  {Nostrum.Struct.Guild.id(),
   old_member :: Nostrum.Struct.Guild.Member.t() | nil,
   updated_member :: Nostrum.Struct.Guild.Member.t()}

Update the given member for the given guild from upstream data.

Return the guild ID that was updated, the old cached member (if the member was known to the cache), and the updated member.

Note regarding intents

Even if the required intents to receive GUILD_MEMBER_UPDATE events are disabled to a point where we do not receive guild creation events, it is still possible to receive the event for our own user. An example of this can be found in issue #293. Note that the linked issue refers to the old contents of this module before the ETS-based guild cache was moved into Elixir.Nostrum.Cache.GuildCache.ETS.

role_create(id, role)

Specs

role_create(Nostrum.Struct.Guild.id(), role :: map()) ::
  Nostrum.Struct.Guild.Role.t()

Create a role on the given guild from upstream data.

Return the casted role.

role_delete(id, id)

Specs

role_delete(Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.id()) ::
  {Nostrum.Struct.Guild.id(), old_role :: Nostrum.Struct.Guild.Role.t()} | :noop

Delete the given role on the given guild.

Return the guild and the old role if it was cached, or :noop otherwise.

role_update(id, role)

Specs

role_update(Nostrum.Struct.Guild.id(), role :: map()) ::
  {old_role :: Nostrum.Struct.Guild.Role.t(),
   new_role :: Nostrum.Struct.Guild.Role.t()}

Update a role on the given guild from upstream data.

Return the old role before the update and the updated role.

select(id, selector)

Specs

select(Nostrum.Struct.Guild.id(), selector()) ::
  {:ok, any()} | {:error, reason()}

Selects values using a selector from a Nostrum.Struct.Guild.

Returns {:error, reason} if no result was found.

Examples

iex> Nostrum.Cache.GuildCache.select(0, fn guild -> guild.id end)
{:ok, 0}

iex> Nostrum.Cache.GuildCache.select(10, fn guild -> guild.id end)
{:error, :id_not_found_on_guild_lookup}

select_all(selector)

Specs

select_all(selector :: (Nostrum.Struct.Guild.t() -> any())) :: Enum.t()

Selects guilds matching selector from all Nostrum.Struct.Guild in the cache.

select_by(clauses, selector)

Specs

select_by(clauses(), selector()) :: {:ok, any()} | {:error, reason()}

Selects values using a selector from a Nostrum.Struct.Guild that matches the clauses.

Returns {:error, reason} if no result was found.

iex> Nostrum.Cache.GuildCache.select_by([id: 0], fn guild -> guild.id end)
{:ok, 0}

iex> Nostrum.Cache.GuildCache.select_by(%{id: 0}, fn guild -> guild.id end)
{:ok, 0}

iex> Nostrum.Cache.GuildCache.select_by([id: 10], fn guild -> guild.id end)
{:error, :id_not_found_on_guild_lookup}

update(map)

Specs

update(map()) ::
  {old_guild :: Nostrum.Struct.Guild.t(),
   updated_guild :: Nostrum.Struct.Guild.t()}

Update a guild from upstream data.

Return the original guild before the update, and the updated guild.

voice_state_update(id, state)

Specs

voice_state_update(Nostrum.Struct.Guild.id(), state :: map()) ::
  {Nostrum.Struct.Guild.id(), new_state :: [map()]}

Update the voice state of the given guild from upstream data.

Note that it is recommended to drop the :member / "member" keys of the supplied upstream data, as these would otherwise duplicate the data that is being kept in the guild cache already.

Return the guild ID and the updated voice states of the guild.

Functions

all()

See Nostrum.Cache.GuildCache.ETS.all/0.

channel_create(guild_id, channel)

See Nostrum.Cache.GuildCache.ETS.channel_create/2.

channel_delete(guild_id, channel_id)

See Nostrum.Cache.GuildCache.ETS.channel_delete/2.

channel_update(guild_id, channel)

See Nostrum.Cache.GuildCache.ETS.channel_update/2.

create(guild)

See Nostrum.Cache.GuildCache.ETS.create/1.

delete(guild_id)

See Nostrum.Cache.GuildCache.ETS.delete/1.

emoji_update(guild_id, emojis)

See Nostrum.Cache.GuildCache.ETS.emoji_update/2.

get(guild_id)

See Nostrum.Cache.GuildCache.ETS.get/1.

get!(guild_id)

Specs

get!(Nostrum.Struct.Guild.id()) :: Nostrum.Struct.Guild.t() | no_return()

Same as get/1, but raises Nostrum.Error.CacheError in case of failure.

get_by(clauses)

See Nostrum.Cache.GuildCache.ETS.get_by/1.

get_by!(clauses)

Specs

get_by!(clauses()) :: Nostrum.Struct.Guild.t() | no_return()

Same as get_by/1, but raises Nostrum.Error.CacheError in case of failure.

member_add(guild_id, member)

See Nostrum.Cache.GuildCache.ETS.member_add/2.

member_chunk(guild_id, member)

See Nostrum.Cache.GuildCache.ETS.member_chunk/2.

member_remove(guild_id, member)

See Nostrum.Cache.GuildCache.ETS.member_remove/2.

member_update(guild_id, member)

See Nostrum.Cache.GuildCache.ETS.member_update/2.

role_create(guild_id, role)

See Nostrum.Cache.GuildCache.ETS.role_create/2.

role_delete(guild_id, role)

See Nostrum.Cache.GuildCache.ETS.role_delete/2.

role_update(guild_id, role)

See Nostrum.Cache.GuildCache.ETS.role_update/2.

select(guild_id, selector)

See Nostrum.Cache.GuildCache.ETS.select/2.

select!(guild_id, selector)

Specs

select!(Nostrum.Struct.Guild.id(), selector()) :: any() | no_return()

Same as select/2, but raises Nostrum.Error.CacheError in case of failure.

select_all(selector)

See Nostrum.Cache.GuildCache.ETS.select_all/1.

select_by(clauses, selector)

See Nostrum.Cache.GuildCache.ETS.select_by/2.

select_by!(clauses, selector)

Specs

select_by!(clauses(), selector()) :: any() | no_return()

Same as select_by/2, but raises Nostrum.Error.CacheError in case of failure.

update(guild)

See Nostrum.Cache.GuildCache.ETS.update/1.

voice_state_update(guild_id, state)

See Nostrum.Cache.GuildCache.ETS.voice_state_update/2.