Nostrum.Cache.GuildCache.ETS (nostrum v0.11.0-dev)

An ETS-based cache for guilds.

The supervisor defined by this module will set up the ETS table associated with it.

The default table name under which guilds are cached is nostrum_guilds. In addition to the cache behaviour implementations provided by this module, you can also call regular ETS table methods on it, such as :ets.info.

Note that users should not call the functions not related to this specific implementation of the cache directly. Instead, call the functions of Nostrum.Cache.GuildCache directly, which will dispatch to the configured cache.

Functions

channel_create(guild_id, channel)
(since 0.5.0)

@spec channel_create(Nostrum.Struct.Guild.id(), map()) :: Nostrum.Struct.Channel.t()

Create the given channel for the given guild in the cache.

channel_delete(guild_id, channel_id)
(since 0.5.0)

@spec channel_delete(Nostrum.Struct.Guild.id(), Nostrum.Struct.Channel.id()) ::
  Nostrum.Struct.Channel.t() | :noop

Delete the channel from the given guild in the cache.

channel_update(guild_id, channel)
(since 0.5.0)

@spec channel_update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Channel.t(), Nostrum.Struct.Channel.t()}

Update the channel on the given guild in the cache.

child_spec(init_arg)
(since 0.5.0)

Returns a specification to start this module under a supervisor.

See Supervisor.

create(payload)
(since 0.5.0)

@spec create(map()) :: Nostrum.Struct.Guild.t()

Create the given guild in the cache.

delete(guild_id)
(since 0.5.0)

@spec delete(Nostrum.Struct.Guild.id()) :: Nostrum.Struct.Guild.t() | nil

Delete the given guild from the cache.

emoji_update(guild_id, emojis)
(since 0.5.0)

@spec emoji_update(Nostrum.Struct.Guild.id(), [map()]) ::
  {[Nostrum.Struct.Emoji.t()], [Nostrum.Struct.Emoji.t()]}

Update the emoji list for the given guild in the cache.

get(guild_id)
(since 0.5.0)

@spec get(Nostrum.Struct.Guild.id()) ::
  {:ok, Nostrum.Struct.Guild.t()} | {:error, :not_found}

Get a guild from the cache.

init(init_arg)
(since 0.5.0)

Set up the cache's ETS table.

member_count_down(guild_id)
(since 0.7.0)

@spec member_count_down(Nostrum.Struct.Guild.id()) :: true

Decrement the guild member count by one.

member_count_up(guild_id)
(since 0.7.0)

@spec member_count_up(Nostrum.Struct.Guild.id()) :: true

Increment the guild member count by one.

role_create(guild_id, role)
(since 0.5.0)

@spec role_create(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.t()}

Create the given role in the given guild in the cache.

role_delete(guild_id, role_id)
(since 0.5.0)

@spec role_delete(Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.id()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.t()} | :noop

Delete the given role from the given guild in the cache.

role_update(guild_id, role)
(since 0.5.0)

@spec role_update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Role.t(),
   Nostrum.Struct.Guild.Role.t()}

Update the given role in the given guild in the cache.

start_link(init_arg)
(since 0.5.0)

Start the supervisor.

stickers_update(guild_id, stickers)
(since 0.5.0)

@spec stickers_update(Nostrum.Struct.Guild.id(), [map()]) ::
  {[Nostrum.Struct.Sticker.t()], [Nostrum.Struct.Sticker.t()]}

Update the sticker list for the given guild in the cache.

tabname()
(since 0.5.0)

@spec tabname() :: atom()

Retrieve the ETS table name used for the cache.

update(payload)
(since 0.5.0)

@spec update(map()) :: {Nostrum.Struct.Guild.t() | nil, Nostrum.Struct.Guild.t()}

Update the given guild in the cache.

voice_state_update(guild_id, payload)
(since 0.5.0)

@spec voice_state_update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), [map()]}

Update guild voice states with the given voice state in the cache.