Nostrum.Cache.MemberCache.ETS (nostrum v0.11.0-dev)

An ETS-based cache for members.

If you need to get the table reference for the table used by this module, please use the table/0 function.

Functions

bulk_create(guild_id, members)
(since 0.7.0)

Bulk create a chunk of members for the given guild in the cache.

by_guild(guild_id)
(since 0.10.0)

Stream all members on the given guild.

by_user(user_id)
(since 0.10.0)

Stream all members with the given user ID.

child_spec(init_arg)
(since 0.7.0)

Returns a specification to start this module under a supervisor.

See Supervisor.

clear()
(since 0.8.0)

@spec clear() :: :ok

Clear any objects in the cache.

create(guild_id, payload)
(since 0.7.0)

@spec create(Nostrum.Struct.Guild.id(), map()) :: Nostrum.Struct.Guild.Member.t()

Add the given member to the given guild in the cache.

delete(guild_id, user_id)
(since 0.7.0)

@spec delete(Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.user_id()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.t()} | :noop

Remove the given member from the given guild in the cache.

get(guild_id, user_id)
(since 0.7.0)

@spec get(Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.user_id()) ::
  {:ok, Nostrum.Struct.Guild.Member.t()} | {:error, atom()}

Retrieve the member for the given guild and user in the cache.

init(init_arg)
(since 0.7.0)

Set up the cache's ETS table.

start_link(init_arg)
(since 0.7.0)

Start the supervisor.

table()
(since 0.8.0)

@spec table() :: :ets.table()

Retrieve the ETS table reference used for the cache.

update(guild_id, payload)
(since 0.7.0)

@spec update(Nostrum.Struct.Guild.id(), map()) ::
  {Nostrum.Struct.Guild.id(), Nostrum.Struct.Guild.Member.t() | nil,
   Nostrum.Struct.Guild.Member.t()}

Update the given member for the given guild in the cache.

wrap_query(fun)
(since 0.10.0)

@spec wrap_query((-> query_result)) :: query_result when query_result: term()

Wrap long-running queries operations.

Safety

Any operations are surrounded by :ets.safe_fixtable. It is therefore recommended to finish your read quickly.