Nostrum.Cache.UserCache behaviour (Nostrum v0.5.0)

Cache behaviour & dispatcher for users.

Callbacks

bulk_create(user_payloads)

Specs

bulk_create(user_payloads :: [Map.t()]) :: :ok

Bulk add multiple users to the cache at once.

Returns :ok.

create(payload)

Specs

create(payload :: Map.t()) :: Nostrum.Struct.User.t()

Add a new user to the cache based on the Discord Gateway payload.

Returns a Nostrum.Struct.User.t/0 struct representing the created user.

delete(snowflake)

Specs

delete(snowflake :: Nostrum.Struct.User.id()) :: :noop | Nostrum.Struct.User.t()

Delete a user by ID.

Returns the deleted user if present in the cache, or :noop if the user was not cached.

get(id)

Specs

get(id :: Nostrum.Struct.User.id()) ::
  {:ok, Nostrum.Struct.User.t()} | {:error, atom()}

Retrieves a user from the cache by id.

If successful, returns {:ok, user}. Otherwise, returns {:error, reason}.

Example

case Nostrum.Cache.UserCache.get(1111222233334444) do
  {:ok, user} ->
    "We found " <> user.username
  {:error, _reason} ->
    "No es bueno"
end

update(payload)

Specs

update(payload :: Map.t()) ::
  :noop | {Nostrum.Struct.User.t(), Nostrum.Struct.User.t()}

Update a user in the cache based on payload sent via the Gateway.

Returns :noop if the user has not been updated in the cache, or {old_user, new_user} is the user has been written to the cache.

Functions

bulk_create(users)

Specs

bulk_create([Map.t()]) :: :ok

Bulk create multiple users using the selected cache implementation.

create(payload)

Specs

create(Map.t()) :: Nostrum.Struct.User.t()

Create a user using the selected cache implementation.

delete(id)

Specs

delete(Nostrum.Struct.User.id()) :: :noop | Nostrum.Struct.User.t()

Delete a user by ID using the selected cache implementation.

get(id)

Specs

get(Nostrum.Struct.User.id()) ::
  {:error, atom()} | {:ok, Nostrum.Struct.User.t()}

Retrieve a user using the selected cache implementation.

get!(id)

Specs

get!(Nostrum.Struct.User.id()) :: no_return() | Nostrum.Struct.User.t()

Same as get/1, but raises Nostrum.Error.CacheError in case of a failure.

update(payload)

Specs

update(Map.t()) :: :noop | {Nostrum.Struct.User.t(), Nostrum.Struct.User.t()}

Update the given user using the selected cache implementation.