Nostrum.Constants.ApplicationCommandOptionType (Nostrum v0.9.0)

Defines available types used for defining application command option types for passed options For more info please refer to https://discord.com/developers/docs/interactions/application-commands#application-command-object-application-command-option-type

Functions

attachment()

boolean()

channel()

integer()

mentionable()

number()

role()

string()

sub_command()

sub_command_group()

user()