Nostrum.StateMachineTranslator (Nostrum v0.10.0)

Translate error reports for :gen_statem modules in Elixir.

Internal module

This module is intended for exclusive usage inside of nostrum, and is documented for completeness and people curious to look behind the covers.

Reasoning

By default, Elixir will ignore these messages altogether, see this ElixirForum post. A possible workaround seems to be using the gen_state_machine library, but pulling in a library purely to have error reporting for something built-in to OTP seems pretty strange to me.

Functions

translate(min_level, arg2, arg3, arg4)
(since 0.9.0)