Nostrum.Struct.ApplicationCommandInteractionDataOption (Nostrum v0.9.0-alpha3)

Struct for command invocation arguments.

Types

focused()
(since 0.5.0)

@type focused() :: boolean() | nil

Whether this parameter is focused for autocomplete interactions.

name()

@type name() :: String.t()

Parameter name

options()

@type options() :: [t()] | nil

Parameter options for subcommands.

Mutually exclusive with value.

t()

@type t() :: %Nostrum.Struct.ApplicationCommandInteractionDataOption{
  focused: focused(),
  name: name(),
  options: options(),
  type: type(),
  value: value()
}

Command interaction data struct

type()
(since 0.5.0)

@type type() :: 1..10

The application command option type.

See https://discord.com/developers/docs/interactions/application-commands#application-command-object-application-command-option-type for more details.

You can use one of the Nostrum.Constants.ApplicationCommandOptionType methods.

value()

@type value() ::
  String.t()
  | integer()
  | boolean()
  | number()
  | Nostrum.Struct.User.id()
  | Nostrum.Struct.Channel.id()
  | Nostrum.Struct.Guild.Role.id()
  | nil

Parameter value.

The type of this depends on the type/0:

For type/0 of 3, this will be a String.t/0.

For type/0 of 4, this will be a integer/0.

For type/0 of 5, this will be a boolean/0.

For type/0 of 6, this will be a Nostrum.Struct.User.id/0. The corresponding guild member and user can be looked up in Nostrum.Struct.ApplicationCommandInteractionData.resolved/0.

For type/0 of 7, this will be a Nostrum.Struct.Channel.id/0. The corresponding channel can be looked up in Nostrum.Struct.ApplicationCommandInteractionData.resolved/0.

For type/0 of 8, this will be a Nostrum.Struct.Guild.Role.id/0. The corresponding role can be looked up in Nostrum.Struct.ApplicationCommandInteractionData.resolved/0.

For type/0 of 9, this will be a Nostrum.Struct.User.id/0 or Nostrum.Struct.Guild.Role.id/0. The corresponding user or role can be looked up in Nostrum.Struct.ApplicationCommandInteractionData.resolved/0.

For type/0 of 10, this will be a number/0.

Mutually exclusive with options. If options is not nil, this will be nil.