Nostrum.Struct.ApplicationCommandInteractionDataResolved (nostrum v0.11.0-dev)

Converted interaction payload.

Types

attachments()
(since 0.7.0)

@type attachments() ::
  %{
    required(Nostrum.Struct.Message.Attachment.id()) =>
      Nostrum.Struct.Message.Attachment.t()
  }
  | nil

Attachments sent along with the interaction.

channels()
(since 0.5.0)

@type channels() ::
  %{
    required(Nostrum.Struct.Channel.id()) =>
      Nostrum.Struct.Channel.guild_text_channel()
  }
  | nil

IDs and corresponding partial channels.

The channels in this map only have the following keys set:

id

name

type

permissions

members()
(since 0.5.0)

@type members() ::
  %{required(Nostrum.Struct.User.id()) => Nostrum.Struct.Guild.Member.t()} | nil

IDs and corresponding partial members.

These members are missing values on the following fields:

user

deaf

mute

The corresponding user data can be looked up in users. For members that are part of this map, data for the corresponding user will always be included.

messages()
(since 0.5.0)

@type messages() ::
  %{required(Nostrum.Struct.Message.id()) => Nostrum.Struct.Message.t()} | nil

The IDs and corresponding messages.

roles()
(since 0.5.0)

@type roles() ::
  %{required(Nostrum.Struct.Guild.Role.id()) => Nostrum.Struct.Guild.Role.t()}
  | nil

IDs and corresponding roles

t()
(since 0.5.0)

@type t() :: %Nostrum.Struct.ApplicationCommandInteractionDataResolved{
  attachments: attachments(),
  channels: channels(),
  members: members(),
  messages: messages(),
  roles: roles(),
  users: users()
}

Resolved interaction data

users()
(since 0.5.0)

@type users() ::
  %{required(Nostrum.Struct.User.id()) => Nostrum.Struct.User.t()} | nil

IDs and corresponding users