Nostrum.Struct.Component.ActionRow (Nostrum v0.10.0)

Action Rows.

Types

t()
(since 0.5.0)

@type t() :: %Nostrum.Struct.Component{
  channel_types: term(),
  components: [Nostrum.Struct.Component.components()],
  custom_id: term(),
  default_values: term(),
  disabled: term(),
  emoji: term(),
  label: term(),
  max_length: term(),
  max_values: term(),
  min_length: term(),
  min_values: term(),
  options: term(),
  placeholder: term(),
  required: term(),
  style: term(),
  type: Nostrum.Struct.Component.type(),
  url: term(),
  value: term()
}

Functions

action_row(opts \\ [])
(since 0.5.0)

Create an empty action row.

Options can be passed as a keyword list. The only supported option is a list of inner components

append(action_row, button)
(since 0.5.0)

Appends a button to the action row.

Returns the action row unchanged if there are already 5 buttons or if the action row contains a select menu or text input.

append_lazy(action_row, button)
(since 0.5.0)

Lazily appends a button to the action row.

If there are already 5 buttons, the first one will be dropped.

flatten(map)
(since 0.5.0)

put(component, select_menu)
(since 0.5.0)

Puts the given component into the action row, any existing components are discarded.

put_new(component, list_of_components)
(since 0.5.0)

Puts the components into the action row unless a list of inner components already exists.