Elixir v0.3.0 Nostrum.Struct.Emoji

Struct representing a Discord emoji.

Using Emojis in Messages

A Nostrum.Struct.Emoji can be used in message content using the String.Chars protocol or format_mention/1. We highly recommend using the String.Chars protocol instead of the latter.

emoji = %Nostrum.Struct.Emoji{id: 43819043108, name: "foxbot"}

Nostrum.Api.create_message!(189098431762321, "#{emoji}")
Nostrum.Api.create_message!(189098431762321, "#{Nostrum.Struct.Emoji.format_mention(emoji)}")

Using Emojis in the Api

See Nostrum.Struct.Emoji.emoji_api_name/0.

Types

animated()

animated() :: boolean() | nil

Whether this emoji is animated

emoji_api_name()

emoji_api_name() :: String.t()

Emoji string to be used with the Discord API.

Some API endpoints take an emoji. If it is a custom emoji, it must be structured as "id:name". If it is an unicode emoji, it can be structured as any of the following:

"name"

A base 16 unicode emoji string.

We suggest library users to use the get_api_name/1 function to get a Nostrum.Struct.Emoji’s emoji api name.

Examples

# Custom Emojis
"nostrum:431890438091489"

# Unicode Emojis
"≡ƒæì"
"\xF0\x9F\x98\x81"
"\u2b50"

# All Emojis
emoji = %Nostrum.Struct.Emoji{id: 43819043108, name: "foxbot"}
Nostrum.Struct.Emoji.get_api_name(emoji)
"foxbot:43819043108"

id()

id() :: Nostrum.Struct.Snowflake.t() | nil

Id of the emoji

managed()

managed() :: boolean() | nil

Whether this emoji is managed

name()

name() :: String.t()

Name of the emoji

require_colons()

require_colons() :: boolean() | nil

Whether this emoji must be wrapped in colons

roles()

roles() :: [Nostrum.Struct.Snowflake.t()] | nil

Roles this emoji is whitelisted to

t()

t() :: %Nostrum.Struct.Emoji{
  animated: animated(),
  id: id(),
  managed: managed(),
  name: name(),
  require_colons: require_colons(),
  roles: roles(),
  user: user()
}

user()

user() :: Nostrum.Struct.User.t() | nil

User that created this emoji

Functions

format_mention(emoji)

format_mention(t()) :: String.t()

Formats an emoji struct into an emoji mention.

Because Nostrum.Struct.Emoji implements the String.Chars protocol, we recommend nostrum users to use String.Chars instead of this function. See Nostrum.Struct.Emoji for more information.

Examples

emoji = %Nostrum.Struct.Emoji{id: 43819043108, name: "foxbot"}

Nostrum.Struct.Emoji.format_mention(emoji)
"<:foxbot:43819043108>"

Nostrum.Api.create_message!(4318940318049, "Sending some text with this emoji #{Nostrum.Struct.Emoji.format_mention(emoji)}")

get_api_name(emoji)

get_api_name(t()) :: emoji_api_name()

Formats an emoji struct into its Nostrum.Struct.Emoji.emoji_api_name/0.

Examples

emoji = %Nostrum.Struct.Emoji{id: 43819043108, name: "foxbot"}

api_name = Nostrum.Struct.Emoji.get_api_name(emoji)
"foxbot:43819043108"

Nostrum.Api.create_reaction(471635274857, 47361625345554, api_name)
{:ok}