Nostrum.Struct.Event.PollVoteChange (nostrum v0.11.0-dev)

Represents an addition or removal of a vote from a Discord poll.

For polls where multiple answers were selected, one of these events will be fired for each vote.

Types

answer_id()

@type answer_id() :: integer()

ID corresponding to the answer_id in the Nostrum.Struct.Message.Poll.answers/0 list

channel_id()

@type channel_id() :: Nostrum.Struct.Channel.id()

ID of the channel the vote took place in

guild_id()

@type guild_id() :: Nostrum.Struct.Guild.id()

ID of the guild the poll is in (unless it is a private channel)

message_id()

@type message_id() :: Nostrum.Struct.Message.id()

ID of the message the poll was attached to

t()

@type t() :: %Nostrum.Struct.Event.PollVoteChange{
  answer_id: answer_id(),
  channel_id: channel_id(),
  guild_id: guild_id(),
  message_id: message_id(),
  type: type(),
  user_id: user_id()
}

Event representing a addition or removal of a vote from a poll

type()

@type type() :: :add | :remove

Whether the vote was an addition or removal for a vote of the option

user_id()

@type user_id() :: Nostrum.Struct.User.id()

ID of the user that has voted