Nostrum.Struct.Event.Ready (nostrum v0.11.0-dev)

Sent after initial handshake with the gateway

Types

application()
(since 0.5.0)

@type application() :: Nostrum.Struct.Event.PartialApplication.t()

Partial application object with id and flags

guilds()
(since 0.5.0)

@type guilds() :: [Nostrum.Struct.Guild.UnavailableGuild.t()]

The guilds that the bot user is in

resume_gateway_url()
(since 0.5.0)

@type resume_gateway_url() :: String.t()

Similar to session_id/0, this is the URL that Discord has requested reconnection attempts to be made against.

Nostrum stores and handles this for you, this value is provided for debugging purposes.

session_id()
(since 0.5.0)

@type session_id() :: String.t()

Used for resuming connections.

If you are wondering whether you need to use this, you probably don't. Nostrum handles reconnections for you.

shard()
(since 0.5.0)

@type shard() :: {integer(), non_neg_integer()} | nil

A pair of two integers {shard_id, num_shards}.

For more information, see https://discord.com/developers/docs/topics/gateway#sharding.

t()
(since 0.5.0)

@type t() :: %Nostrum.Struct.Event.Ready{
  application: application(),
  guilds: guilds(),
  resume_gateway_url: resume_gateway_url(),
  session_id: session_id(),
  shard: shard(),
  user: user(),
  v: v()
}

Event sent after initial handshake with the gateway

user()
(since 0.5.0)

@type user() :: Nostrum.Struct.User.t()

Information about the bot user

v()
(since 0.5.0)

@type v() :: non_neg_integer()

Gateway version. See https://discord.com/developers/docs/topics/gateway#gateways-gateway-versions