Nostrum.Struct.Event.SpeakingUpdate (Nostrum v0.4.6)

Struct representing a Nostrum-generated Speaking Update event

Nostrum will generate this event when the bot starts or stops playing audio.

Types

channel_id()

Specs

channel_id() :: Nostrum.Struct.Channel.id()

Id of the channel this speaking update is occurring in.

guild_id()

Specs

guild_id() :: Nostrum.Struct.Guild.id()

Id of the guild this speaking update is occurring in.

speaking()

Specs

speaking() :: boolean()

Boolean representing if bot has started or stopped speaking.

t()

Specs

t() :: %Nostrum.Struct.Event.SpeakingUpdate{
  channel_id: channel_id(),
  guild_id: guild_id(),
  speaking: speaking()
}