Nostrum.Struct.Event.ThreadListSync (Nostrum v0.10.0)

Struct representing a Thread List Sync event.

This event is sent when the user gains access to a channel.

Types

channel_ids()
(since 0.5.1)

@type channel_ids() :: [Nostrum.Struct.Channel.id()] | nil

The parent channel ids whose threads are being synced. If omitted, all threads were synced for the entire guild.

guid_id()
(since 0.5.1)

@type guid_id() :: Nostrum.Struct.Guild.id()

The id of the guild.

members()
(since 0.5.1)

@type members() :: [Nostrum.Struct.ThreadMember.t()]

All thread member objects from the synced threads for the current user, indicating which threads the user has been added to.

t()
(since 0.5.1)

@type t() :: %Nostrum.Struct.Event.ThreadListSync{
  channel_ids: channel_ids(),
  guild_id: guid_id(),
  members: members(),
  threads: threads()
}

threads()
(since 0.5.1)

@type threads() :: [Nostrum.Struct.Channel.t()]

All active threads in the given channels that the user can access.