Nostrum.Struct.Event.ThreadMembersUpdate (nostrum v0.11.0-dev)

Struct representing a thread members update event.

This event is sent whenever a user is added or removed from a thread.

If the current user does not have the GUILD_MEMBERS intent, this event will only be sent when the current user is added to or removed from a thread.

Types

added_members()
(since 0.5.1)

@type added_members() :: [Nostrum.Struct.ThreadMember.t()] | nil

The members that were added to the thread.

guild_id()
(since 0.5.1)

@type guild_id() :: Nostrum.Struct.Guild.id()

The id of the guild the thread is in.

id()
(since 0.5.1)

@type id() :: Nostrum.Struct.Channel.id()

The id of the thread.

member_count()
(since 0.5.1)

@type member_count() :: non_neg_integer()

The approximate number of members in the thread.

This number is capped at 50, though there can be more members in the thread.

removed_member_ids()
(since 0.5.1)

@type removed_member_ids() :: [Nostrum.Struct.User.id()] | nil

The ids of the members that were removed from the thread.

t()
(since 0.5.1)

@type t() :: %Nostrum.Struct.Event.ThreadMembersUpdate{
  added_members: added_members(),
  guild_id: guild_id(),
  id: id(),
  member_count: member_count(),
  removed_member_ids: removed_member_ids()
}