Elixir v0.3.0 Nostrum.Struct.Guild

Struct representing a Discord guild.

Types

afk_channel_id()

afk_channel_id() :: Nostrum.Struct.Snowflake.t() | nil

The id of the guild’s afk channel

afk_timeout()

afk_timeout() :: integer()

The time someone must be afk before being moved

application_id()

application_id() :: Nostrum.Struct.Snowflake.t() | nil

Application id of the guild creator if it is bot created.

available_guild()

available_guild() :: %Nostrum.Struct.Guild{
  afk_channel_id: afk_channel_id(),
  afk_timeout: afk_timeout(),
  application_id: application_id(),
  channels: channels(),
  default_message_notifications: default_message_notifications(),
  embed_channel_id: embed_channel_id(),
  embed_enabled: embed_enabled(),
  emojis: emojis(),
  explicit_content_filter: explicit_content_filter(),
  features: features(),
  icon: icon(),
  id: id(),
  joined_at: joined_at(),
  large: large(),
  member_count: member_count(),
  members: members(),
  mfa_level: mfa_level(),
  name: name(),
  owner_id: owner_id(),
  presences: presences(),
  region: region(),
  roles: roles(),
  splash: splash(),
  system_channel_id: system_channel_id(),
  unavailable: false,
  verification_level: verification_level(),
  voice_states: voice_states(),
  widget_channel_id: widget_channel_id(),
  widget_enabled: widget_enabled()
}

A Nostrum.Struct.Guild that is fully available.

channels()

channels() :: [Nostrum.Struct.Channel.t()] | nil

List of channels

default_message_notifications()

default_message_notifications() :: integer()

Default message notifications level.

embed_channel_id()

embed_channel_id() :: Nostrum.Struct.Snowflake.t() | nil

The id of the embedded channel

embed_enabled()

embed_enabled() :: boolean() | nil

Whether the guild is emeddable

emojis()

emojis() :: [Nostrum.Struct.Emoji.t()]

List of emojis

explicit_content_filter()

explicit_content_filter() :: integer()

Explicit content filter level.

features()

features() :: [String.t()]

List of guild features

icon()

icon() :: String.t() | nil

The hash of the guild’s icon

id()

id() :: Nostrum.Struct.Snowflake.t()

The guild’s id

joined_at()

joined_at() :: String.t() | nil

Date the bot user joined the guild

large()

large() :: boolean() | nil

Whether the guild is considered ‘large’

member_count()

member_count() :: integer() | nil

Total number of members in the guild

members()

members() :: [Nostrum.Struct.Guild.Member.t()] | nil

List of members

mfa_level()

mfa_level() :: integer()

Required MFA level of the guild

name()

name() :: String.t()

The name of the guild.

owner_id()

owner_id() :: Nostrum.Struct.Snowflake.t()

The id of the guild owner

presences()

presences() :: [map()]

List of simple presence maps

region()

region() :: String.t()

The id of the voice region

rest_guild()

rest_guild() :: %Nostrum.Struct.Guild{
  afk_channel_id: afk_channel_id(),
  afk_timeout: afk_timeout(),
  application_id: application_id(),
  channels: nil,
  default_message_notifications: default_message_notifications(),
  embed_channel_id: embed_channel_id(),
  embed_enabled: embed_enabled(),
  emojis: emojis(),
  explicit_content_filter: explicit_content_filter(),
  features: features(),
  icon: icon(),
  id: id(),
  joined_at: nil,
  large: nil,
  member_count: nil,
  members: nil,
  mfa_level: mfa_level(),
  name: name(),
  owner_id: owner_id(),
  presences: nil,
  region: region(),
  roles: roles(),
  splash: splash(),
  system_channel_id: system_channel_id(),
  unavailable: nil,
  verification_level: verification_level(),
  voice_states: nil,
  widget_channel_id: widget_channel_id(),
  widget_enabled: widget_enabled()
}

A Nostrum.Struct.Guild that is sent on guild-specific rest endpoints.

roles()

roles() :: [Nostrum.Struct.Guild.Role.t()]

List of roles

splash()

splash() :: String.t() | nil

The hash of the guild’s splash

system_channel_id()

system_channel_id() :: Nostrum.Struct.Snowflake.t() | nil

The id of the channel to which system messages are sent.

t()

t() :: available_guild() | unavailable_guild() | rest_guild() | user_guild()

unavailable()

unavailable() :: boolean() | nil

Whether the guild is avaliable

unavailable_guild()

unavailable_guild() :: %Nostrum.Struct.Guild{
  afk_channel_id: nil,
  afk_timeout: nil,
  application_id: nil,
  channels: nil,
  default_message_notifications: nil,
  embed_channel_id: nil,
  embed_enabled: nil,
  emojis: nil,
  explicit_content_filter: nil,
  features: nil,
  icon: nil,
  id: id(),
  joined_at: nil,
  large: nil,
  member_count: nil,
  members: nil,
  mfa_level: nil,
  name: nil,
  owner_id: nil,
  presences: nil,
  region: nil,
  roles: nil,
  splash: nil,
  system_channel_id: nil,
  unavailable: true,
  verification_level: nil,
  voice_states: nil,
  widget_channel_id: nil,
  widget_enabled: nil
}

A Nostrum.Struct.Guild that is unavailable.

user_guild()

user_guild() :: %Nostrum.Struct.Guild{
  afk_channel_id: nil,
  afk_timeout: nil,
  application_id: nil,
  channels: nil,
  default_message_notifications: nil,
  embed_channel_id: nil,
  embed_enabled: nil,
  emojis: nil,
  explicit_content_filter: nil,
  features: nil,
  icon: icon(),
  id: id(),
  joined_at: nil,
  large: nil,
  member_count: nil,
  members: nil,
  mfa_level: nil,
  name: name(),
  owner_id: nil,
  presences: nil,
  region: nil,
  roles: nil,
  splash: nil,
  system_channel_id: nil,
  unavailable: nil,
  verification_level: nil,
  voice_states: nil,
  widget_channel_id: nil,
  widget_enabled: nil
}

A Nostrum.Struct.Guild that is sent on user-specific rest endpoints.

verification_level()

verification_level() :: integer()

The level of verification

voice_states()

voice_states() :: [map()] | nil

List of voice states as maps

widget_channel_id()

widget_channel_id() :: Nostrum.Struct.Snowflake.t()

The channel id for the server widget.

widget_enabled()

widget_enabled() :: boolean() | nil

Whether or not the server widget is enabled.

Functions

icon_url(guild, image_format \\ "webp")

icon_url(t(), String.t()) :: String.t() | nil

Returns the URL of a guild’s icon, or nil if there is no icon.

Supported image formats are PNG, JPEG, and WebP.

Examples

iex> guild = %Nostrum.Struct.Guild{icon: "86e39f7ae3307e811784e2ffd11a7310",
...>                               id: 41771983423143937}
iex> Nostrum.Struct.Guild.icon_url(guild)
"https://cdn.discordapp.com/icons/41771983423143937/86e39f7ae3307e811784e2ffd11a7310.webp"
iex> Nostrum.Struct.Guild.icon_url(guild, "png")
"https://cdn.discordapp.com/icons/41771983423143937/86e39f7ae3307e811784e2ffd11a7310.png"

iex> guild = %Nostrum.Struct.Guild{icon: nil}
iex> Nostrum.Struct.Guild.icon_url(guild)
nil

splash_url(guild, image_format \\ "webp")

splash_url(t(), String.t()) :: String.t() | nil

Returns the URL of a guild’s splash, or nil if there is no splash.

Supported image formats are PNG, JPEG, and WebP.

Examples

iex> guild = %Nostrum.Struct.Guild{splash: "86e39f7ae3307e811784e2ffd11a7310",
...>                               id: 41771983423143937}
iex> Nostrum.Struct.Guild.splash_url(guild)
"https://cdn.discordapp.com/splashes/41771983423143937/86e39f7ae3307e811784e2ffd11a7310.webp"
iex> Nostrum.Struct.Guild.splash_url(guild, "png")
"https://cdn.discordapp.com/splashes/41771983423143937/86e39f7ae3307e811784e2ffd11a7310.png"

iex> guild = %Nostrum.Struct.Guild{splash: nil}
iex> Nostrum.Struct.Guild.splash_url(guild)
nil