Elixir v0.3.0 Nostrum.Struct.Guild.Member

Struct representing a Discord guild member.

A Nostrum.Struct.Guild.Member stores a Nostrum.Struct.User’s properties pertaining to a specific Nostrum.Struct.Guild.

Mentioning Members in Messages

A Nostrum.Struct.Guild.Member can be mentioned in message content using the String.Chars protocol or mention/1.

member = %Nostrum.Struct.Guild.Member{user: Nostrum.Struct.User{id: 120571255635181568}}
Nostrum.Api.create_message!(184046599834435585, "#{member}")
%Nostrum.Struct.Message{content: "<@120571255635181568>"}

member = %Nostrum.Struct.Guild.Member{user: Nostrum.Struct.User{id: 89918932789497856}}
Nostrum.Api.create_message!(280085880452939778, "#{Nostrum.Struct.Guild.Member.mention(member)}")
%Nostrum.Struct.Message{content: "<@89918932789497856>"}

Types

deaf()

deaf() :: boolean()

Whether the user is deafened

joined_at()

joined_at() :: String.t()

Date the user joined the guild

mute()

mute() :: boolean()

Whether the user is muted

nick()

nick() :: String.t() | nil

The nickname of the user

roles()

roles() :: [Nostrum.Struct.Snowflake.t()]

A list of role ids

t()

t() :: %Nostrum.Struct.Guild.Member{
  deaf: deaf(),
  joined_at: joined_at(),
  mute: mute(),
  nick: nick(),
  roles: roles(),
  user: user()
}

user()

user() :: Nostrum.Struct.User.t()

The user struct

Functions

guild_channel_permissions(member, guild, channel_id)

guild_channel_permissions(
  t(),
  Nostrum.Struct.Guild.t(),
  Nostrum.Struct.Channel.id()
) :: [Nostrum.Struct.Permission.t()]

Returns a member’s permissions in a guild channel, based on its Nostrum.Struct.Overwrites.

Examples

guild = Nostrum.Cache.GuildCache.get!(279093381723062272)
member = Enum.find(guild.members, & &1.id === 177888205536886784)
channel_id = 381889573426429952
Nostrum.Struct.Guild.Member.guild_channel_permissions(member, guild, channel_id)
#=> [:manage_messages]

guild_permissions(member, guild)

guild_permissions(t(), Nostrum.Struct.Guild.t()) :: [
  Nostrum.Struct.Permission.t()
]

Returns a member’s guild permissions.

Examples

guild = Nostrum.Cache.GuildCache.get!(279093381723062272)
member = Enum.find(guild.members, & &1.id === 177888205536886784)
Nostrum.Struct.Guild.Member.guild_permissions(member, guild)
#=> [:administrator]

mention(member)

mention(t()) :: String.t()

Formats a Nostrum.Struct.Guild.Member into a mention.

Examples

iex> member = %Nostrum.Struct.Guild.Member{user: %Nostrum.Struct.User{id: 177888205536886784}}
...> Nostrum.Struct.Guild.Member.mention(member)
"<@177888205536886784>"