Elixir v0.3.0 Nostrum.Struct.Guild.Member

Struct representing a Discord guild member.

User vs. Member

A user contains only general information about that user such as a username and an avatar. A member has everything that a user has, but also additional information on a per guild basis. This includes things like a nickname and a list of roles.

Types

deaf()

deaf() :: boolean()

Whether the user is deafened

joined_at()

joined_at() :: String.t()

Date the user joined the guild

mute()

mute() :: boolean()

Whether the user is muted

nick()

nick() :: String.t() | nil

The nickname of the user

roles()

roles() :: [integer()]

A list of role ids

t()

t() :: %Nostrum.Struct.Guild.Member{
  deaf: deaf(),
  joined_at: joined_at(),
  mute: mute(),
  nick: nick(),
  roles: roles(),
  user: user()
}

user()

user() :: Nostrum.Struct.User.t()

The user struct