Nostrum.Struct.Guild.Role (Nostrum v0.4.6)

Struct representing a Discord role.

Mentioning Roles in Messages

A Nostrum.Struct.Guild.Role can be mentioned in message content using the String.Chars protocol or mention/1.

role = %Nostrum.Struct.Guild.Role{id: 431886897539973131}
Nostrum.Api.create_message!(184046599834435585, "#{role}")
%Nostrum.Struct.Message{}

role = %Nostrum.Struct.Guild.Role{id: 431884023535632398}
Nostrum.Api.create_message!(280085880452939778, "#{Nostrum.Struct.Guild.Role.mention(role)}")
%Nostrum.Struct.Message{}

Types

color()

Specs

color() :: integer()

The hexadecimal color code

hoist()

Specs

hoist() :: boolean()

Whether the role is pinned in the user listing

id()

Specs

id() :: Nostrum.Snowflake.t()

The id of the role

managed()

Specs

managed() :: boolean()

Whether the role is managed by an integration

mentionable()

Specs

mentionable() :: boolean()

Whether the role is mentionable

name()

Specs

name() :: String.t()

The name of the role

permissions()

Specs

permissions() :: integer()

The permission bit set

position()

Specs

position() :: integer()

The position of the role

t()

Specs

t() :: %Nostrum.Struct.Guild.Role{
  color: color(),
  hoist: hoist(),
  id: id(),
  managed: managed(),
  mentionable: mentionable(),
  name: name(),
  permissions: permissions(),
  position: position()
}

Functions

mention(role)

Specs

mention(t()) :: String.t()

Formats an Nostrum.Struct.Role into a mention.

Examples

iex> role = %Nostrum.Struct.Guild.Role{id: 431886639627763722}
...> Nostrum.Struct.Guild.Role.mention(role)
"<@&431886639627763722>"