Nostrum.Struct.Guild.ScheduledEvent (Nostrum v0.10.0)

Struct representing a scheduled event in a guild.

Types

channel_id()

@type channel_id() :: Nostrum.Snowflake.t() | nil

The id of the channel the scheduled event is in. Will be nil if entity_type is EXTERNAL

creator()

@type creator() :: Nostrum.Struct.User.t() | nil

The user who created the scheduled event. Only present when retrieving the event from the API.

note: This will be nil for events created before October 25th, 2021.

creator_id()

@type creator_id() :: Nostrum.Snowflake.t() | nil

The id of the user who created the scheduled event.

note: This will be nil only for events created before October 25th, 2021.

description()

@type description() :: String.t() | nil

The description of the scheduled event.

entity_id()

@type entity_id() :: Nostrum.Snowflake.t() | nil

the id of an entity associated with a guild scheduled event.

entity_metadata()

@type entity_metadata() ::
  Nostrum.Struct.Guild.ScheduledEvent.EntityMetadata.t() | nil

Holds additional metadata associated with a scheduled event.

entity_type()

@type entity_type() :: 1..3

The type of entity the scheduled event is for.

1 - STAGE_INSTANCE 2 - VOICE 3 - EXTERNAL

guild_id()

@type guild_id() :: Nostrum.Snowflake.t()

The id of the guild the scheduled event is in.

id()

@type id() :: Nostrum.Snowflake.t()

The id of the scheduled event.

name()

@type name() :: String.t()

The name of the scheduled event.

privacy_level()

@type privacy_level() :: 2

The privacy level of the scheduled event. At the time of writing, this is always 2 for GUILD_ONLY

scheduled_end_time()

@type scheduled_end_time() :: DateTime.t() | nil

The time the scheduled event ends as an ISO8601 timestamp. Only required if entity_type is EXTERNAL

scheduled_start_time()

@type scheduled_start_time() :: DateTime.t()

The time the scheduled event starts.

status()

@type status() :: 1..4

The status of the scheduled event.

1 - SCHEDULED 2 - ACTIVE 3 - COMPLETED 4 - CANCELLED

note: Once status is set to Completed or Cancelled it can no longer be updated.

t()

@type t() :: %Nostrum.Struct.Guild.ScheduledEvent{
  channel_id: channel_id(),
  creator: creator(),
  creator_id: creator_id(),
  description: description(),
  entity_id: entity_id(),
  entity_metadata: entity_metadata(),
  entity_type: entity_type(),
  guild_id: guild_id(),
  id: id(),
  name: name(),
  privacy_level: privacy_level(),
  scheduled_end_time: scheduled_end_time(),
  scheduled_start_time: scheduled_start_time(),
  status: status(),
  user_count: user_count()
}

user_count()

@type user_count() :: non_neg_integer() | nil

The number of users who have subscribed to the scheduled event.

Functions

to_struct(map)