Nostrum.Struct.Invite (nostrum v0.11.0-dev)

Struct representing a Discord invite.

Types

approximate_member_count()

@type approximate_member_count() :: integer() | nil

Approximate total member count of the guild this invite is for.

approximate_presence_count()

@type approximate_presence_count() :: integer() | nil

Approximate online member count of the guild this invite is for.

channel()

@type channel() :: Nostrum.Struct.Channel.t()

The channel this invite is for.

code()

@type code() :: String.t()

The invite code (unique ID).

created_at()

@type created_at() :: String.t()

When this invite was created.

detailed_invite()

@type detailed_invite() :: %Nostrum.Struct.Invite{
  approximate_member_count: approximate_member_count(),
  approximate_presence_count: approximate_presence_count(),
  channel: channel(),
  code: code(),
  created_at: created_at(),
  guild: guild(),
  inviter: inviter(),
  max_age: max_age(),
  max_uses: max_uses(),
  target_user: target_user(),
  target_user_type: target_user_type(),
  temporary: temporary(),
  uses: uses()
}

An invite with metadata.

guild()

@type guild() :: Nostrum.Struct.Guild.t() | nil

Partially populated guild struct of the guild this invite is for.

inviter()

@type inviter() :: Nostrum.Struct.User.t() | nil

The user who created this invite.

max_age()

@type max_age() :: integer()

Duration (in seconds) after which the invite expires.

max_uses()

@type max_uses() :: integer()

Max number of times this invite can be used.

simple_invite()

@type simple_invite() :: %Nostrum.Struct.Invite{
  approximate_member_count: approximate_member_count(),
  approximate_presence_count: approximate_presence_count(),
  channel: channel(),
  code: code(),
  created_at: nil,
  guild: guild(),
  inviter: inviter(),
  max_age: nil,
  max_uses: nil,
  target_user: target_user(),
  target_user_type: target_user_type(),
  temporary: nil,
  uses: nil
}

An invite without metadata.

t()

@type t() :: simple_invite() | detailed_invite()

target_user()

@type target_user() :: Nostrum.Struct.User.t() | nil

Partially populated user struct of the target user for this invite.

target_user_type()

@type target_user_type() :: integer() | nil

The type of user target for this invite.

temporary()

@type temporary() :: boolean()

Whether this invite only grants temporary membership.

uses()

@type uses() :: integer()

Number of times this invite has been used.