Nostrum.Struct.Message (Nostrum v0.4.6)

Struct representing a Discord message.

Types

activity()

Specs

activity() :: Nostrum.Struct.Message.Activity.t() | nil

The activity of the message. Sent with Rich Presence-related chat embeds.

application()

Specs

application() :: Nostrum.Struct.Message.Application.t() | nil

The application of the message. Sent with Rich Presence-related chat embeds.

attachments()

Specs

attachments() :: [Nostrum.Struct.Message.Attachment.t()]

List of attached files in the message

author()

Specs

author() :: Nostrum.Struct.User.t()

The user struct of the author

channel_id()

Specs

channel_id() :: Nostrum.Struct.Channel.id()

The id of the channel

content()

Specs

content() :: String.t()

The content of the message

edited_timestamp()

Specs

edited_timestamp() :: String.t() | nil

When the message was edited

embeds()

Specs

embeds() :: [Nostrum.Struct.Embed.t()]

List of embedded content in the message

guild_id()

Specs

guild_id() :: Nostrum.Struct.Guild.id() | nil

The id of the guild

id()

Specs

id() :: Nostrum.Snowflake.t()

The id of the message

member()

Specs

member() :: Nostrum.Struct.Guild.Member.t() | nil

Partial Guild Member object received with the Message Create event if message came from a guild channel.

mention_everyone()

Specs

mention_everyone() :: boolean()

Whether this messsage mentions everyone

mention_roles()

Specs

mention_roles() :: [Nostrum.Struct.Guild.Role.id()]

List of roles ids mentioned in the message

mentions()

Specs

mentions() :: [Nostrum.Struct.User.t()]

List of users mentioned in the message

message_reference()

Specs

message_reference() :: Nostrum.Struct.Message.Reference.t() | nil

Reference data sent with crossposted messages and replies

nonce()

Specs

nonce() :: String.t() | nil

Validates if a message was sent

pinned()

Specs

pinned() :: boolean()

Whether this message is pinned

reactions()

Specs

reactions() :: [Nostrum.Struct.Message.Reaction.t()] | nil

Reactions to the message.

referenced_message()

Specs

referenced_message() :: t() | nil

The message associated with the message_reference

This field is only returned for messages with a type of 19 (REPLY). If the message is a reply but the referenced_message field is not present, the backend did not attempt to fetch the message that was being replied to, so its state is unknown. If the field exists but is null, the referenced message was deleted.

t()

Specs

t() :: %Nostrum.Struct.Message{
  activity: activity(),
  application: application(),
  attachments: attachments(),
  author: author(),
  channel_id: channel_id(),
  content: content(),
  edited_timestamp: edited_timestamp(),
  embeds: embeds(),
  guild_id: guild_id(),
  id: id(),
  member: member(),
  mention_everyone: mention_everyone(),
  mention_roles: mention_roles(),
  mentions: mentions(),
  message_reference: message_reference(),
  nonce: nonce(),
  pinned: pinned(),
  reactions: reactions(),
  referenced_message: referenced_message(),
  timestamp: timestamp(),
  tts: tts(),
  type: type(),
  webhook_id: webhook_id()
}

timestamp()

Specs

timestamp() :: String.t()

When the message was sent

tts()

Specs

tts() :: boolean()

Whether this was a TTS message

type()

Specs

type() :: integer()

Type of message.

webhook_id()

Specs

webhook_id() :: Nostrum.Snowflake.t() | nil

If the message is generated by a webhook, this is the webhook's id.