Nostrum.Struct.ThreadMember (Nostrum v0.9.0-alpha3)

Struct representing a thread member object

Types

flags()
(since 0.5.1)

@type flags() :: non_neg_integer()

Any user-thread settings

guild_id()
(since 0.5.1)

@type guild_id() :: Nostrum.Struct.Guild.id() | nil

The id of the guild the thread is in.

Only present within THREAD_MEMBER_UPDATE events

id()
(since 0.5.1)

@type id() :: Nostrum.Snowflake.t() | nil

The id of the thread, omitted within GUILD_CREATE events

join_timestamp()
(since 0.5.1)

@type join_timestamp() :: DateTime.t()

The timestamp of when the user last joined the thread

t()
(since 0.5.1)

@type t() :: %Nostrum.Struct.ThreadMember{
  flags: flags(),
  guild_id: guild_id(),
  id: id(),
  join_timestamp: join_timestamp(),
  user_id: user_id()
}

user_id()
(since 0.5.1)

@type user_id() :: Nostrum.Struct.User.id() | nil

The id of the user, omitted within GUILD_CREATE events