Nostrum.Struct.User.Flags (Nostrum v0.4.5)

Struct representing the flags a user account can have

Types

bug_hunter_level_1()

Specs

bug_hunter_level_1() :: bool()

Bug Hunter (Level 1)

bug_hunter_level_2()

Specs

bug_hunter_level_2() :: bool()

Bug Hunter (Level 2)

early_supporter()

Specs

early_supporter() :: bool()

Early Supporter

flags()

Specs

flags() :: %Nostrum.Struct.User.Flags{
  bug_hunter_level_1: bug_hunter_level_1(),
  bug_hunter_level_2: bug_hunter_level_2(),
  early_supporter: early_supporter(),
  hypesquad_balance: hypesquad_balance(),
  hypesquad_bravery: hypesquad_bravery(),
  hypesquad_brilliance: hypesquad_brilliance(),
  hypesquad_events: hypesquad_events(),
  partner: partner(),
  staff: staff(),
  system: system(),
  team_user: team_user(),
  verified_bot: verified_bot(),
  verified_developer: verified_developer()
}

hypesquad_balance()

Specs

hypesquad_balance() :: bool()

HypeSquad Balance

hypesquad_bravery()

Specs

hypesquad_bravery() :: bool()

HypeSquad Bravery

hypesquad_brilliance()

Specs

hypesquad_brilliance() :: bool()

HypeSquad Brilliance

hypesquad_events()

Specs

hypesquad_events() :: bool()

HypeSquad Events

partner()

Specs

partner() :: bool()

Discord Partner

staff()

Specs

staff() :: bool()

Discord Employee

system()

Specs

system() :: bool()

System user

t()

Specs

t() :: flags()

team_user()

Specs

team_user() :: bool()

Team User

verified_bot()

Specs

verified_bot() :: bool()

Verified bot

verified_developer()

Specs

verified_developer() :: bool()

Verified developer

Functions

from_integer(flag_value)

Specs

from_integer(integer()) :: t()

Constructs a flag struct based on an integer from the Discord API (either public_flags or flags).

Examples

iex> Nostrum.Struct.User.Flags.from_integer(131842)
%Nostrum.Struct.User.Flags{
  bug_hunter_level_1: false,
  bug_hunter_level_2: false,
  early_supporter: true,
  hypesquad_balance: true,
  hypesquad_bravery: false,
  hypesquad_brilliance: false,
  hypesquad_events: false,
  partner: true,
  staff: false,
  system: false,
  team_user: false,
  verified_bot: false,
  verified_developer: true
}

to_integer(flag_struct)

Specs

to_integer(t()) :: integer()

Convert a flag struct to an integer value.

Examples

iex> my_flags = %Nostrum.Struct.User.Flags{
...>  bug_hunter_level_1: false,
...>  bug_hunter_level_2: false,
...>  early_supporter: true,
...>  hypesquad_balance: true,
...>  hypesquad_bravery: false,
...>  hypesquad_brilliance: false,
...>  hypesquad_events: false,
...>  partner: true,
...>  staff: false,
...>  system: false,
...>  team_user: false,
...>  verified_bot: false,
...>  verified_developer: true
...> }
iex> Nostrum.Struct.User.Flags.to_integer(my_flags)
131842