Nostrum.Struct.VoiceWSState (Nostrum v0.5.0)

Struct representing the current Voice WS state.

Types

conn()

Specs

conn() :: pid()

PID of the :gun worker connected to the websocket

conn_pid()

Specs

conn_pid() :: pid()

PID of the connection process

gateway()

Specs

gateway() :: String.t()

Gateway URL

guild_id()

Specs

guild_id() :: Nostrum.Struct.Guild.id()

The guild id that this voice websocket state applies to

heartbeat_ack()

Specs

heartbeat_ack() :: boolean()

Whether or not the last hearbeat sent was ACK'd

heartbeat_interval()

Specs

heartbeat_interval() :: integer() | nil

Interval at which heartbeats are sent

heartbeat_ref()

Specs

heartbeat_ref() :: :timer.tref() | nil

Time ref for the heartbeat

identified()

Specs

identified() :: boolean()

Whether the session has been identified

last_heartbeat_ack()

Specs

last_heartbeat_ack() :: DateTime.t() | nil

The time the last heartbeat was acknowledged, will be nil if a heartbeat hasn't been ACK'd yet

last_heartbeat_send()

Specs

last_heartbeat_send() :: DateTime.t()

The time the last heartbeat was sent, if a heartbeat hasn't been sent it will be the time the websocket process was started

session()

Specs

session() :: String.t()

The session id

stream()

Specs

stream() :: :gun.stream_ref()

Stream reference for :gun

t()

Specs

t() :: %Nostrum.Struct.VoiceWSState{
  conn: conn(),
  conn_pid: conn_pid(),
  gateway: gateway(),
  guild_id: guild_id(),
  heartbeat_ack: heartbeat_ack(),
  heartbeat_interval: heartbeat_interval(),
  heartbeat_ref: heartbeat_ref(),
  identified: identified(),
  last_heartbeat_ack: last_heartbeat_ack(),
  last_heartbeat_send: last_heartbeat_send(),
  session: session(),
  stream: stream(),
  token: token()
}

token()

Specs

token() :: String.t()

The session token