Nostrum.Struct.WSState (Nostrum v0.4.6)

Struct representing the current WS state.

Types

conn()

Specs

conn() :: pid()

PID of the :gun worker connected to the websocket

conn_pid()

Specs

conn_pid() :: pid()

PID of the connection process

gateway()

Specs

gateway() :: String.t()

Gateway URL

heartbeat_ack()

Specs

heartbeat_ack() :: boolean()

Whether or not the last hearbeat sent was ACK'd

heartbeat_interval()

Specs

heartbeat_interval() :: integer() | nil

Interval at which heartbeats are sent

heartbeat_ref()

Specs

heartbeat_ref() :: integer() | nil

Time ref for the heartbeat

last_heartbeat_ack()

Specs

last_heartbeat_ack() :: DateTime.t() | nil

The time the last heartbeat was acknowledged, will be nil if a heartbeat hasn't been ACK'd yet

last_heartbeat_send()

Specs

last_heartbeat_send() :: DateTime.t()

The time the last heartbeat was sent, if a heartbeat hasn't been sent it will be the time the websocket process was started

seq()

Specs

seq() :: integer() | nil

The sequence number of the last event

session()

Specs

session() :: integer() | nil

The session id

shard_num()

Specs

shard_num() :: non_neg_integer()

The shard number

shard_pid()

Specs

shard_pid() :: pid()

PID of the shard containing this state

t()

Specs

t() :: %Nostrum.Struct.WSState{
  conn: conn(),
  conn_pid: conn_pid(),
  gateway: gateway(),
  heartbeat_ack: heartbeat_ack(),
  heartbeat_interval: heartbeat_interval(),
  heartbeat_ref: heartbeat_ref(),
  last_heartbeat_ack: last_heartbeat_ack(),
  last_heartbeat_send: last_heartbeat_send(),
  seq: seq(),
  session: session(),
  shard_num: shard_num(),
  shard_pid: shard_pid(),
  zlib_ctx: zlib_ctx()
}

zlib_ctx()

Specs

zlib_ctx() :: reference() | nil

Reference to the current zlib context